% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logging_helpers.R
\name{add_base_logger_appenders}
\alias{add_base_logger_appenders}
\title{Add base logger appenders for top-level setup logs
This is for backwards compatibility now that logs are moved into a new folder for
 new gpa objects}
\usage{
add_base_logger_appenders(
  lg,
  gpa,
  log_txt_path,
  log_json_path,
  txt_appender_name,
  json_appender_name
)
}
\arguments{
\item{lg}{A logger instance}

\item{gpa}{glm_pipeline_arguments object containing logging settings}

\item{log_txt_path}{Path to the text log file}

\item{log_json_path}{Path to the json log file}

\item{txt_appender_name}{Appender name for text logs}

\item{json_appender_name}{Appender name for json logs}
}
\description{
Add base logger appenders for top-level setup logs
This is for backwards compatibility now that logs are moved into a new folder for
 new gpa objects
}
\keyword{internal}
