% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmri_utility_fx.R
\name{add_baseline_regressors}
\alias{add_baseline_regressors}
\title{Helper function to append baseline regressors to the convolved design matrix.
  These are especially useful for GLMs that use concatenated data for multiple runs,
  where run-specific intercepts and drift terms are needed.}
\usage{
add_baseline_regressors(
  dmat_convolved,
  baseline_coef_order = -1L,
  baseline_parameterization = "Legendre"
)
}
\arguments{
\item{dmat_convolved}{The design matrix containing convolved regressors for all events,
created by \code{build_design_matrix}}

\item{baseline_coef_order}{The polynomial order of baseline regressors to be added}

\item{baseline_parameterization}{The parameterization of baseline regressors.
Default is "Legendre". Alternative is "orthogonal.polynomials".}
}
\value{
A modified version of \code{dmat_convolved} that contains the baseline regressors
  requested by the user.
}
\description{
Helper function to append baseline regressors to the convolved design matrix.
  These are especially useful for GLMs that use concatenated data for multiple runs,
  where run-specific intercepts and drift terms are needed.
}
\keyword{internal}
