% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fsl_helper_functions.R
\name{add_custom_feat_syntax}
\alias{add_custom_feat_syntax}
\title{Helper function to find/replace or insert additional arguments to include
  in FSF syntax for a FEAT model.}
\usage{
add_custom_feat_syntax(fsf_syntax, feat_args, lg = NULL)
}
\arguments{
\item{fsf_syntax}{a character vector containing all fsf syntax}

\item{feat_args}{a named list}
}
\value{
a modified version of \code{fsf_syntax} that now includes all custom
  arguments from \code{feat_args}
}
\description{
Helper function to find/replace or insert additional arguments to include
  in FSF syntax for a FEAT model.
}
\details{
If a tag already exists in the fsf syntax, it is replaced by the custom argument
  in \code{feat_args}. If the tag does not exist, it is inserted.
}
\keyword{internal}
