% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/job_tracking_functions.R
\name{add_tracked_job_parent}
\alias{add_tracked_job_parent}
\title{Add parent/child id relationship to tracking database}
\usage{
add_tracked_job_parent(
  sqlite_db = NULL,
  job_id = NULL,
  parent_job_id = NULL,
  child_level = 1
)
}
\arguments{
\item{sqlite_db}{Path to SQLite database used for tracking}

\item{job_id}{Job id of job for which to add a parent}

\item{parent_job_id}{Job id of the parent job to job_id}

\item{child_level}{Level of child; currently supports two levels}
}
\description{
Add parent/child id relationship to tracking database
}
