% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/afni_whereami.R
\name{afni_whereami}
\alias{afni_whereami}
\title{wrapper class for AFNI whereami}
\description{
wrapper class for AFNI whereami
}
\section{Methods}{
\subsection{Public methods}{
  \itemize{
    \item \href{#method-afni_whereami-initialize}{\code{afni_whereami$new()}}
    \item \href{#method-afni_whereami-get_call}{\code{afni_whereami$get_call()}}
    \item \href{#method-afni_whereami-get_omask_call}{\code{afni_whereami$get_omask_call()}}
    \item \href{#method-afni_whereami-get_whereami_df}{\code{afni_whereami$get_whereami_df()}}
    \item \href{#method-afni_whereami-get_overlap_df}{\code{afni_whereami$get_overlap_df()}}
    \item \href{#method-afni_whereami-get_omask_df}{\code{afni_whereami$get_omask_df()}}
    \item \href{#method-afni_whereami-get_outputs}{\code{afni_whereami$get_outputs()}}
    \item \href{#method-afni_whereami-run}{\code{afni_whereami$run()}}
    \item \href{#method-afni_whereami-clone}{\code{afni_whereami$clone()}}
  }
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-afni_whereami-initialize"></a>}}
\if{latex}{\out{\hypertarget{method-afni_whereami-initialize}{}}}
\subsection{\code{afni_whereami$new()}}{
  create new afni_whereami object
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{afni_whereami$new(
  afni_3dclusterize_obj = NULL,
  omask = NULL,
  coord_file = NULL,
  coord_file_columns = NULL,
  coord_vector = NULL,
  atlases = NULL,
  coord_orientation = NULL,
  coord_space = NULL,
  output_file = NULL,
  omask_output_file = NULL,
  afnidir = NULL
)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{afni_3dclusterize_obj}}{Use an existing 3dClusterize object to build the whereami lookup}
      \item{\code{omask}}{Use this integer-valued cluster mask to lookup percent overlap with different atlas parcels}
      \item{\code{coord_file}}{The text file containing coordinates for the clusters to lookup}
      \item{\code{coord_file_columns}}{The columns in \code{coord_file} containing the x, y, and z coordinates.
Note that these should be zero-based indices, so the first column is 0, second column is 1, etc.}
      \item{\code{coord_vector}}{Rather than lookup multiple locations from a file, \code{coord_vector} requires
three numeric values (x, y, and z coordinates) that are used to lookup a single region in the atlases.}
      \item{\code{atlases}}{a character vector of atlases in whereami to use for lookup}
      \item{\code{coord_orientation}}{LPI or RAI}
      \item{\code{coord_space}}{The template space for the coordinates. Default: 'MNI'}
      \item{\code{output_file}}{The file name/path for the output of whereami}
      \item{\code{omask_output_file}}{The file name/path for the output of whereami with the omask (percentage overlap)}
      \item{\code{afnidir}}{The location of afni}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-afni_whereami-get_call"></a>}}
\if{latex}{\out{\hypertarget{method-afni_whereami-get_call}{}}}
\subsection{\code{afni_whereami$get_call()}}{
  return the whereami call for this object
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{afni_whereami$get_call()}
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-afni_whereami-get_omask_call"></a>}}
\if{latex}{\out{\hypertarget{method-afni_whereami-get_omask_call}{}}}
\subsection{\code{afni_whereami$get_omask_call()}}{
  return the whereami call for the omask lookup
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{afni_whereami$get_omask_call()}
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-afni_whereami-get_whereami_df"></a>}}
\if{latex}{\out{\hypertarget{method-afni_whereami-get_whereami_df}{}}}
\subsection{\code{afni_whereami$get_whereami_df()}}{
  return the data.frame of coordinates and labels from the whereami lookup
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{afni_whereami$get_whereami_df()}
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-afni_whereami-get_overlap_df"></a>}}
\if{latex}{\out{\hypertarget{method-afni_whereami-get_overlap_df}{}}}
\subsection{\code{afni_whereami$get_overlap_df()}}{
  return the data.frame of percent overlap between each atlas and clusters/ROIs in the input mask
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{afni_whereami$get_overlap_df()}
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-afni_whereami-get_omask_df"></a>}}
\if{latex}{\out{\hypertarget{method-afni_whereami-get_omask_df}{}}}
\subsection{\code{afni_whereami$get_omask_df()}}{
  return the data.frame of the percentage overlap for each parcel (not implemented)
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{afni_whereami$get_omask_df()}
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-afni_whereami-get_outputs"></a>}}
\if{latex}{\out{\hypertarget{method-afni_whereami-get_outputs}{}}}
\subsection{\code{afni_whereami$get_outputs()}}{
  return a list of the output files from this whereami object
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{afni_whereami$get_outputs(exclude_missing = TRUE)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{exclude_missing}}{if TRUE, any expected output file that does not exist will be NA in the returned list}
    }
    \if{html}{\out{</div>}}
  }
  \subsection{Returns}{
    a list of all output files
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-afni_whereami-run"></a>}}
\if{latex}{\out{\hypertarget{method-afni_whereami-run}{}}}
\subsection{\code{afni_whereami$run()}}{
  Run whereami for this object
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{afni_whereami$run(force = FALSE)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{force}}{if TRUE, re-run a whereami call even if the expected output file exists}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-afni_whereami-clone"></a>}}
\if{latex}{\out{\hypertarget{method-afni_whereami-clone}{}}}
\subsection{\code{afni_whereami$clone()}}{
  The objects of this class are cloneable with this method.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{afni_whereami$clone(deep = FALSE)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{deep}}{Whether to make a deep clone.}
    }
    \if{html}{\out{</div>}}
  }
}

}
