% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/l1_helper_functions.R
\name{as.character.formula}
\alias{as.character.formula}
\title{S3 method to convert formula to character}
\usage{
\method{as.character}{formula}(x, ...)
}
\arguments{
\item{x}{a formula to convert to character}

\item{...}{not used}
}
\description{
S3 method to convert formula to character
}
\details{
copied from formula.tools to avoid NAMESPACE import
}
