% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glm_helper_functions.R
\name{as_dhms}
\alias{as_dhms}
\title{helper function to convert a period object back into a d-h:m:s string}
\usage{
as_dhms(per)
}
\description{
helper function to convert a period object back into a d-h:m:s string
}
\keyword{internal}
