% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_l1_models.R
\name{bl1_build_models}
\alias{bl1_build_models}
\title{helper function to create a set of level 1 models}
\usage{
bl1_build_models(l1_model_set, spec_list = NULL, lg = NULL)
}
\arguments{
\item{l1_model_set}{a set of level 1 models}

\item{spec_list}{if creating models from a spec (YAML) file, this is the parsed list}

\item{lg}{the current logger}
}
\description{
helper function to create a set of level 1 models
}
\keyword{internal}
