% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_l1_models.R
\name{bl1_get_cols}
\alias{bl1_get_cols}
\title{Onset, duration, value column selection helper function}
\usage{
bl1_get_cols(
  l1_model_set,
  trial_data,
  field_name = NULL,
  field_desc = NULL,
  select_cols = NULL,
  select_regex = NULL,
  force_selection = TRUE
)
}
\arguments{
\item{l1_model_set}{an l1_model_set object containing onsets etc.}

\item{trial_data}{a trial + subjects x events data.frame that contains potential onset columns}

\item{field_name}{the element of \code{l1_model_set} containing columns of a certain purpose (onset, duration, value)}

\item{field_desc}{the text description of the field being modified (e.g., 'parametric value')}

\item{select_cols}{a character vector of current columns specified by the user to be added/included}

\item{select_regex}{a PCRE-compatible regular expression for identifying columns}
}
\value{
a modified version of l1_model_set that has \code{field_name} updated according to user specification
}
\description{
Onset, duration, value column selection helper function
}
\keyword{internal}
