% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_fwe_correction.R
\name{build_3dclustsim_permutation}
\alias{build_3dclustsim_permutation}
\title{helper function for setting up 3dClustSim objects for permutation (via 3dttest++).}
\usage{
build_3dclustsim_permutation(to_fwe)
}
\arguments{
\item{to_fwe}{a data.frame containing model information for the group analyses to be passed through 3dClustSim FWE}
}
\description{
helper function for setting up 3dClustSim objects for permutation (via 3dttest++).
}
\keyword{internal}
