% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/afni_3dlmer.R
\name{build_3dlmer_datatable}
\alias{build_3dlmer_datatable}
\title{Internal function to build the data table for AFNI 3dLMEr}
\usage{
build_3dlmer_datatable(subject_data, input_files, model_variables)
}
\arguments{
\item{subject_data}{a data.frame containing subject-level covariates}

\item{input_files}{a data.frame with columns 'id', 'session', and 'InputFile'}

\item{model_variables}{a character vector of variables that must be included in the table}
}
\value{
a data.frame formatted for 3dLMEr -dataTable
}
\description{
Internal function to build the data table for AFNI 3dLMEr
}
\keyword{internal}
