% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_glm_betas_in_mask.R
\name{build_beta_extraction}
\alias{build_beta_extraction}
\title{helper function to prompt user to choose combinations of l1, l2, and l3 models for beta extraction.}
\usage{
build_beta_extraction(
  gpa,
  extract_l1 = "prompt",
  extract_l2 = "prompt",
  extract_l3 = "prompt",
  lg = NULL
)
}
\arguments{
\item{gpa}{a \code{glm_pipeline_arguments} object}

\item{lg}{a logger object for messages}
}
\value{
a list of data frames (l1, l2, l3) consisting of the chosen combinations of l1, l2, and l3 models
}
\description{
helper function to prompt user to choose combinations of l1, l2, and l3 models for beta extraction.
}
\keyword{internal}
