% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_fwe_correction.R
\name{build_fwe_correction}
\alias{build_fwe_correction}
\title{Function to walk user through setting up FWE corrections for model outputs}
\usage{
build_fwe_correction(gpa, lg = NULL)
}
\arguments{
\item{gpa}{a \code{glm_pipeline_arguments} object containing model outputs in l3_model_setup}

\item{lg}{a logger object for logging messages}
}
\value{
returns a modified gpa object that contains FWE specifications in $fwe_correction
}
\description{
Function to walk user through setting up FWE corrections for model outputs
}
\details{
currently only looks a level 3
}
