% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_l1_models.R
\name{build_l1_models}
\alias{build_l1_models}
\title{Interactive function to build an l1 model specification for setup_glm_pipeline}
\usage{
build_l1_models(
  gpa = NULL,
  trial_data = NULL,
  l1_model_set = NULL,
  onset_cols = NULL,
  onset_regex = ".*(onset|time).*",
  duration_regex = ".*duration.*",
  value_cols = NULL
)
}
\arguments{
\item{gpa}{a \code{glm_pipeline_arguments} object containing an analysis pipeline to which $l1_models
shoudl be added. If $l1_models is already present, these will be amended.}

\item{trial_data}{a data.frame containing trial-level data for one or more subjects}

\item{l1_model_set}{optional existing l1_model_set to be modified}

\item{onset_cols}{an optional character vector of columns in \code{trial_data} that should be
in the set of event onsets}

\item{onset_regex}{an optional PCRE-compatible regular expression for identifying potential
event onset columns in \code{trial_data}}

\item{duration_regex}{an optional PCRE-compatible regular expression for identifying potential
event duration columns in \code{trial_data}}

\item{value_cols}{an optional character vector of columns in \code{trial_data} that should be in the set of signal values}
}
\value{
a \code{l1_model_set} object containing events, signals, and models, compatible with build_design_matrix
}
\description{
Interactive function to build an l1 model specification for setup_glm_pipeline
}
\details{
if \code{gpa} is not passed in, then we will work from trial_data and l1_model_set.
}
\author{
Michael Hallquist
}
