% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_l2_models.R
\name{build_l2_models}
\alias{build_l2_models}
\title{Interactive function to build an l2 model specification for setup_glm_pipeline}
\usage{
build_l2_models(gpa, from_spec_file = NULL, regressor_cols = NULL)
}
\arguments{
\item{gpa}{a \code{glm_pipeline_arguments} object containing an analysis pipeline to which $l2_models
should be added. If $l2_models is already present, these will be amended.}

\item{from_spec_file}{optional YAML or JSON file containing settings to populated into l2 models}

\item{regressor_cols}{an optional character vector of columns in \code{data} that should be considered
as possible regressors}
}
\value{
a \code{hi_model_set} object containing a list of higher-level regression models
}
\description{
Interactive function to build an l2 model specification for setup_glm_pipeline
}
\author{
Michael Hallquist
}
