% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_l2_models.R
\name{build_l3_models}
\alias{build_l3_models}
\title{Interactive function to build an l3 model specification for setup_glm_pipeline}
\usage{
build_l3_models(gpa, from_spec_file = NULL, regressor_cols = NULL)
}
\description{
Interactive function to build an l3 model specification for setup_glm_pipeline
}
