% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_motion_exclusions.R
\name{calculate_motion_exclusions}
\alias{calculate_motion_exclusions}
\title{Function to calculate exclusions of runs and subjects based on excessive head movement}
\usage{
calculate_motion_exclusions(gpa)
}
\arguments{
\item{gpa}{a \code{glm_pipeline_arguments} object, as specified by glm_setup_pipeline}
}
\value{
a modified version of the \code{gpa} object that contains a $motion_exclusions data.frame
}
\description{
Function to calculate exclusions of runs and subjects based on excessive head movement
}
\author{
Michael Hallquist
}
\keyword{internal}
