% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glm_helper_functions.R
\name{calculate_subject_exclusions}
\alias{calculate_subject_exclusions}
\title{small helper function to populate subject exclusions in $run_data and $subject_data}
\usage{
calculate_subject_exclusions(gpa)
}
\arguments{
\item{gpa}{a \code{glm_pipeline_arguments} object that already has the \code{$exclude_run}
column populated in \code{$l1_model_setup}.}
}
\value{
a modified copy of \code{gpa} where $exclude_subject has been added to $run_data
  and $subject_data. This function also adds $n_good_runs to $subject_data which is useful
  if we want to enforce a lower bound on the number of runs used to define subject exclusion.
}
\description{
small helper function to populate subject exclusions in $run_data and $subject_data
}
\author{
Michael Hallquist
}
\keyword{internal}
