% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glm_helper_functions.R
\name{check_nums}
\alias{check_nums}
\title{helper function to validate numbers in an input string or vector}
\usage{
check_nums(inp, lower = 0, upper = 1e+10, as_string = TRUE)
}
\arguments{
\item{inp}{A string, character vector, or numeric vector}

\item{lower}{The lowest valid value for each number}

\item{upper}{The highest valid value for each number}

\item{as_string}{If TRUE, return numbers as a space-separated string. If FALSE,
return the validated numeric vector itself.}
}
\description{
helper function to validate numbers in an input string or vector
}
\keyword{internal}
