% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_glm_pipeline.R
\name{choose_glm_models}
\alias{choose_glm_models}
\title{helper function to guide user through process of choosing which models to run in GLM pipeline}
\usage{
choose_glm_models(
  gpa,
  l1_model_names = NULL,
  l2_model_names = NULL,
  l3_model_names = NULL,
  lg = NULL
)
}
\arguments{
\item{l1_model_names}{a character vector of level 1 model names}

\item{l2_model_names}{a character vector of level 2 model names}

\item{l3_model_names}{a character vector of level 3 model names}
}
\value{
a named list containing all models that were selected along with additional information
  about whether to rerun existing models
}
\description{
helper function to guide user through process of choosing which models to run in GLM pipeline
}
\keyword{internal}
