% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glm_helper_functions.R
\name{cleanup_glm_pipeline}
\alias{cleanup_glm_pipeline}
\title{helper function to refresh l3 model status and save gpa object from batch pipeline back to its cache}
\usage{
cleanup_glm_pipeline(gpa)
}
\arguments{
\item{gpa}{a glm_pipeline_arguments object}
}
\value{
a refreshed version of the gpa object
}
\description{
helper function to refresh l3 model status and save gpa object from batch pipeline back to its cache
}
