% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmri_utility_fx.R
\name{cleanup_regressor}
\alias{cleanup_regressor}
\title{Small utility function to handle any NAs in the regressor specification.
Also, optionally, removes values of 0 from the regressor prior to convolution.
The latter is usually a good idea to avoid a non-event (zero) becoming an event
after mean centering of a parametric regressor. Convolving a zero-height regressor
with an HRF will maintain all zeros, whereas a non-zero value (resulting from mean centering)
will not.}
\usage{
cleanup_regressor(times, durations, values, rm_zeros = TRUE)
}
\arguments{
\item{times}{Vector of times (in seconds) for regressor events}

\item{durations}{Vector of durations (in seconds) for regressor events}

\item{values}{Vector of values (parametric heights) for regressor events}

\item{rm_zeros}{Logical indicating whether to remove zero-valued events from regressor}
}
\description{
Small utility function to handle any NAs in the regressor specification.
Also, optionally, removes values of 0 from the regressor prior to convolution.
The latter is usually a good idea to avoid a non-event (zero) becoming an event
after mean centering of a parametric regressor. Convolving a zero-height regressor
with an HRF will maintain all zeros, whereas a non-zero value (resulting from mean centering)
will not.
}
\author{
Michael Hallquist
}
\keyword{internal}
