% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_job_submit.R
\name{cluster_job_submit}
\alias{cluster_job_submit}
\title{This function submits a single script to a high-performance cluster using a scheduler (Slurm or TORQUE).
It accepts a vector of arguments to be passed to the scheduler and
a vector of environment variables that should be passed to the compute node at job execution.}
\usage{
cluster_job_submit(
  script,
  scheduler = "slurm",
  sched_args = NULL,
  env_variables = NULL,
  export_all = FALSE,
  echo = TRUE,
  fail_on_error = FALSE,
  wait_jobs = NULL,
  wait_signal = "afterok",
  repolling_interval = 60,
  tracking_sqlite_db = NULL,
  tracking_args = list()
)
}
\arguments{
\item{script}{A script that should be executed by the scheduler. This can contain scheduler directives, but in the
case of conflicts, the directives passed with \code{sched_args} will take precedence.}

\item{scheduler}{Which scheduler to use for job submission. Options are 'qsub', 'torque', 'sbatch', 'slurm', or 'sh'.
The terms 'qsub' and 'torque' are aliases (where 'torque' submits via the qsub command). Likewise for 'sbatch'
and 'slurm'. The scheduler 'sh' does not submit to any scheduler at all, but instead executes the command
immediately via sh.}

\item{sched_args}{A character vector of arguments to be included in the scheduling command. On TORQUE, these
will typically begin with '-l' such as '-l wall_time=10:00:00'.}

\item{env_variables}{A named character vector containing environment variables and their values to be passed
to the \code{script} at execution time. This is handled by the '-v' directive on TORQUE clusters and
by '--export' on Slurm clusters. The names of this vector are the environment variable names and
the values of the vector are the environment variable values to be passed in.
If you want to propagate the current value of an environment variable to the compute node at runtime,
use NA as the value of the element in \code{env_variables}. See examples.}

\item{export_all}{Whether to export all environment variables to the compute node at runtime. Default: FALSE}

\item{echo}{Whether to echo the job submission command to the terminal at the time it is scheduled. Default: TRUE.}

\item{fail_on_error}{Whether to stop execution of the script (TRUE), or issue a warning (FALSE) if the job
submission fails. Defaults to FALSE (i.e., issue a warning).}

\item{wait_jobs}{a character string of jobs or process ids that should complete before this job is executed}

\item{wait_signal}{on torque or slurm clusters, the signal that should indicate that parent jobs have finished.}

\item{repolling_interval}{The number of seconds to wait before rechecking job status (used only for local scheduler)}

\item{tracking_sqlite_db}{Optional SQLite database used for job tracking.}

\item{tracking_args}{Optional job-tracking metadata passed through to the tracking database.}
}
\value{
A character string containing the jobid of the scheduled job.
}
\description{
The function returns the jobid of the scheduled job.
}
\examples{
\dontrun{
  #simple PBS submission
  cluster_job_submit(
    "myscript.bash",
    scheduler = "torque",
    sched_args = c("-l walltime=10:00:00", "-l nodes=1:ppn=20"),
    env_variables = c(RUN_INDEX = 2, MODEL_NAME = "FSE21")
  )

  #To forward environment variables without explicitly providing values. Note that these must
  #  be in R's system environment (cf. Sys.getenv) at execution time to forward as expected.
  cluster_job_submit('myscript.sbatch', scheduler="slurm",
     sched_args=c('-p general', '-N 1', '-n 12', '--mem=10g', '-t 02-00:00:00'),
     env_variables=c(RUN_INDEX=2, R_HOME=NA, JAVA_HOME=NA))
}

}
\author{
Michael Hallquist
}
