% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_job_submit.R
\name{cluster_submit_shell_jobs}
\alias{cluster_submit_shell_jobs}
\title{helper function to submit a set of shell jobs that are independent of one another}
\usage{
cluster_submit_shell_jobs(
  job_list,
  commands_per_cpu = 1L,
  cpus_per_job = 8L,
  memgb_per_command = 8,
  time_per_job = "1:00:00",
  fork_jobs = TRUE,
  pre = NULL,
  post = NULL,
  sched_args = NULL,
  env_variables = NULL,
  wait_jobs = NULL,
  scheduler = "slurm",
  job_out_dir = getwd(),
  job_script_prefix = "job",
  log_file = "cluster_submit_jobs.csv",
  debug = FALSE
)
}
\arguments{
\item{job_list}{a list or character vector where each element represents an independent job to execute in a shell environment}

\item{commands_per_cpu}{how many elements from \code{job_list} are executed by each core within a single job}

\item{cpus_per_job}{how many cpus/cores are requested for each job}

\item{memgb_per_command}{amount of memory (RAM) requested for each command (in GB)}

\item{time_per_job}{amount of time requested for each job}

\item{job_out_dir}{the directory where job scripts should be written}

\item{job_script_prefix}{the filename prefix for each job script}

\item{log_file}{a csv log file containing the job ids and commands that were executed}

\item{debug}{a logical indicating whether to actually submit the jobs (TRUE) or just create the scripts for inspection (FALSE)}
}
\description{
helper function to submit a set of shell jobs that are independent of one another
}
