% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/afni_clusterize.R
\name{clusterize_spec}
\alias{clusterize_spec}
\title{wrapper class for 3dClusterize}
\description{
wrapper class for 3dClusterize

wrapper class for 3dClusterize
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{fwe_p}}{a vector of p-values used for familywise error (FWE) z-statistic threshold calculations in pTFCE.}
}
\if{html}{\out{</div>}}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{fwe_p}}{a vector of p-values used for familywise error (FWE) z-statistic threshold calculations in pTFCE.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{clusterize_spec$new()}}
\item \href{#method-run}{\code{clusterize_spec$run()}}
\item \href{#method-get_clust_df}{\code{clusterize_spec$get_clust_df()}}
\item \href{#method-get_clust_nifti}{\code{clusterize_spec$get_clust_nifti()}}
\item \href{#method-get_call}{\code{clusterize_spec$get_call()}}
\item \href{#method-clone}{\code{clusterize_spec$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{clusterize_spec$new(
  inset = NULL,
  mask = NULL,
  threshold_file = NULL,
  data_file = NULL,
  mask_from_hdr = NULL,
  out_mask = NULL,
  ithr = NULL,
  idat = NULL,
  onesided = NULL,
  twosided = NULL,
  bisided = NULL,
  lower_thresh = NULL,
  upper_thresh = NULL,
  one_thresh = NULL,
  one_tail = NULL,
  NN = NULL,
  clust_nvox = NULL,
  clust_nvol = NULL,
  pref_map = NULL,
  pref_dat = NULL,
  x1Dformat = NULL,
  no_1Dformat = NULL,
  quiet = NULL,
  orient = NULL,
  binary = NULL,
  clusterize_output_file = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{inset}}{A 4D dataset containing the statistic to use for thresholding (ithr) and, optionally, the data value to output/retain}

\item{\code{mask}}{If specified, the volume will be masked by \code{mask} prior to clusterizing}

\item{\code{threshold_file}}{A 3D dataset containing the statistic to use for thresholding. Mutually exclusive with \code{inset}
If passed, \code{ithr} and \code{idat} are ignored because the \code{inset} file is generated internally.}

\item{\code{data_file}}{A 3D dataset containing the data value to be retained in clusters post-thresholding. 
Must be passed with \code{threshold_file} and will be stitched together with it internally. Mutually exclusive with \code{inset}.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-run"></a>}}
\if{latex}{\out{\hypertarget{method-run}{}}}
\subsection{Method \code{run()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{clusterize_spec$run()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_clust_df"></a>}}
\if{latex}{\out{\hypertarget{method-get_clust_df}{}}}
\subsection{Method \code{get_clust_df()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{clusterize_spec$get_clust_df()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_clust_nifti"></a>}}
\if{latex}{\out{\hypertarget{method-get_clust_nifti}{}}}
\subsection{Method \code{get_clust_nifti()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{clusterize_spec$get_clust_nifti()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_call"></a>}}
\if{latex}{\out{\hypertarget{method-get_call}{}}}
\subsection{Method \code{get_call()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{clusterize_spec$get_call()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{clusterize_spec$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
