% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/afni_clustsim.R
\name{clustsim_spec}
\alias{clustsim_spec}
\title{R6 class for 3dClustSim automation}
\description{
R6 class for 3dClustSim automation

R6 class for 3dClustSim automation
}
\keyword{internal}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{fwhmx_set}}{A fwhmx_set_spec object containing 3dFWHMx information for all fwhmx_input_files}

\item{\code{inset_files}}{A character vector of files to use directly as volumes to threshold and clusterize}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{clustsim_spec$new()}}
\item \href{#method-run}{\code{clustsim_spec$run()}}
\item \href{#method-get_clustsim_df}{\code{clustsim_spec$get_clustsim_df()}}
\item \href{#method-get_call}{\code{clustsim_spec$get_call()}}
\item \href{#method-get_out_dir}{\code{clustsim_spec$get_out_dir()}}
\item \href{#method-get_ncpus}{\code{clustsim_spec$get_ncpus()}}
\item \href{#method-get_use_fwhmx_acf}{\code{clustsim_spec$get_use_fwhmx_acf()}}
\item \href{#method-clone}{\code{clustsim_spec$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{clustsim_spec$new(
  out_dir = NULL,
  prefix = NULL,
  fwhmx_input_files = NULL,
  fwhmx_mask_files = NULL,
  inset_files = NULL,
  dxyz = NULL,
  nxyz = NULL,
  clustsim_mask = NULL,
  acf_params = NULL,
  nopad = NULL,
  pthr = NULL,
  athr = NULL,
  iter = NULL,
  nodec = NULL,
  seed = NULL,
  scheduler = NULL,
  ncpus = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-run"></a>}}
\if{latex}{\out{\hypertarget{method-run}{}}}
\subsection{Method \code{run()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{clustsim_spec$run(force = FALSE)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_clustsim_df"></a>}}
\if{latex}{\out{\hypertarget{method-get_clustsim_df}{}}}
\subsection{Method \code{get_clustsim_df()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{clustsim_spec$get_clustsim_df()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_call"></a>}}
\if{latex}{\out{\hypertarget{method-get_call}{}}}
\subsection{Method \code{get_call()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{clustsim_spec$get_call()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_out_dir"></a>}}
\if{latex}{\out{\hypertarget{method-get_out_dir}{}}}
\subsection{Method \code{get_out_dir()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{clustsim_spec$get_out_dir()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_ncpus"></a>}}
\if{latex}{\out{\hypertarget{method-get_ncpus}{}}}
\subsection{Method \code{get_ncpus()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{clustsim_spec$get_ncpus()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_use_fwhmx_acf"></a>}}
\if{latex}{\out{\hypertarget{method-get_use_fwhmx_acf}{}}}
\subsection{Method \code{get_use_fwhmx_acf()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{clustsim_spec$get_use_fwhmx_acf()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{clustsim_spec$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
