% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spm_utils.R
\name{combine_spm_l3_to_afni}
\alias{combine_spm_l3_to_afni}
\title{Combine SPM L3 outputs into AFNI BRIK+HEAD files for visualization}
\usage{
combine_spm_l3_to_afni(
  gpa,
  spm_l3_combined_filename = NULL,
  spm_l3_combined_briknames = NULL,
  template_brain = NULL
)
}
\arguments{
\item{gpa}{a glm_pipeline_arguments object having a populated l3_model_setup field.}

\item{spm_l3_combined_filename}{a glue expression for the path and filename prefix.}

\item{spm_l3_combined_briknames}{a glue expression for naming the subbriks in the AFNI output.}

\item{template_brain}{an optional filename for the MNI template that should be used as an underlay in AFNI.}
}
\value{
A data.frame describing the images that were combined.
}
\description{
Combine SPM L3 outputs into AFNI BRIK+HEAD files for visualization
}
\details{
This function mirrors the logic of combine_feat_l3_to_afni, but for SPM L3 outputs. It detects
  contrast maps (con_*.nii) and statistic maps (spmT_*.nii / spmF_*.nii) and concatenates them into AFNI
  BRIK/HEAD files with informative labels.
}
