% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glm_helper_functions.R
\name{compute_spike_regressors}
\alias{compute_spike_regressors}
\title{compute spike regressors from a data.frame or matrix of motion parameters and a set of
  expressions that are evaluated against this parameter matrix.}
\usage{
compute_spike_regressors(mot = NULL, spike_volume = NULL, lg = NULL)
}
\arguments{
\item{mot}{a data.frame or matrix with volumes on rows and named motion parameters on columns}

\item{spike_volume}{a character vector of expressions to evaluate against \code{mot}. Resulting
columns will be prefixed with the names of each expression. If the expressions are unnamed,
prefixes will be expr1_, expr2_, etc.}
}
\value{
a matrix of spike regressors (volumes on rows, spike regressors on columns)
}
\description{
compute spike regressors from a data.frame or matrix of motion parameters and a set of
  expressions that are evaluated against this parameter matrix.
}
\keyword{internal}
