% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmri_utility_fx.R
\name{concat_design_runs}
\alias{concat_design_runs}
\title{Concatenate design matrices for each run to form a single design with unique baselines per run (ala AFNI)}
\usage{
concat_design_runs(d, convolved = TRUE)
}
\arguments{
\item{d}{A design matrix object created by \code{build_design_matrix}. The $design_convolved element will be used for concatenation.}

\item{convolved}{If \code{TRUE} (the default), concatenate the convolved design matrix. If \code{FALSE}, use the unconvolved.}
}
\description{
Concatenate design matrices for each run to form a single design with unique baselines per run (ala AFNI)
}
