% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deconvolve_nlreg_resample.R
\name{convolve_dcv_hrf}
\alias{convolve_dcv_hrf}
\title{This function takes in a matrix of generated neural events and
parameters describing the HRF function and convolves the neural
events and HRF function neural events}
\usage{
convolve_dcv_hrf(NEVgen, kernel)
}
\arguments{
\item{NEVgen}{ROIs x time matrix of true neural events generated by the model}

\item{kernel}{The HRF kernel used for convolution}
}
\description{
This function takes in a matrix of generated neural events and
parameters describing the HRF function and convolves the neural
events and HRF function neural events
}
\author{
Keith Bush
}
\keyword{internal}
