% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmri_utility_fx.R
\name{convolve_regressor}
\alias{convolve_regressor}
\title{This function convolves a regressor with a normalized HRF whose peak amplitude is 1.0}
\usage{
convolve_regressor(
  n_vols,
  reg,
  tr = 1,
  normalization = "none",
  rm_zeros = TRUE,
  center_values = TRUE,
  convmax_1 = FALSE,
  demean_convolved = FALSE,
  high_pass = NULL,
  convolve = TRUE,
  beta_series = FALSE,
  ts_multiplier = NULL,
  drop_volumes = 0,
  hrf_parameters = c(a1 = 6, a2 = 12, b1 = 0.9, b2 = 0.9, cc = 0.35)
)
}
\arguments{
\item{n_vols}{The number of volumes (scans) to be output in the convolved regressor}

\item{reg}{A matrix containing the trial, onset, duration, and value for each event}

\item{tr}{The repetition time in seconds}

\item{normalization}{The HRF normalization method used: "none", "durmax_1", or "evtmax_1"}

\item{rm_zeros}{Whether to remove zeros from events vector prior to convolution. Generally a good idea since we typically center
values prior to convolution, and retaining zeros will lead them to be non-zero after mean centering.}

\item{center_values}{Whether to demean values vector before convolution. Default \code{TRUE}.}

\item{convmax_1}{Whether to rescale the convolved regressor to a maximum height of 1.}

\item{demean_convolved}{Whether to demean the regressor after convolution (default: \code{TRUE})}

\item{high_pass}{The cutoff frequency (in Hz) used for high-pass filtering. If \code{NULL}, no filtering is applied.}

\item{convolve}{If \code{TRUE}, the regressor is convolved with the HRF. If \code{FALSE},
the regressor values are simply aligned onto the time grid without convolution
based on the corresponding onsets, durations, and values.}

\item{beta_series}{If \code{TRUE}, split \code{reg} into separate regressors for each event (row). These can be used
to estimate separate betas in the GLM for each event.}

\item{ts_multiplier}{A vector that is n_vols in length that will be multiplied against the stimulus vector before convolution.}

\item{drop_volumes}{The number of volumes to drop from the beginning of the regressor}

\item{hrf_parameters.}{A named vector of parameters passed to \code{fmri.stimulus} that control the shape of the double gamma HRF.
Default: \code{c(a1 = 6, a2 = 12, b1 = 0.9, b2 = 0.9, cc = 0.35)}.}
}
\description{
It extends \code{fmri.stimulus} by allowing for two normalization approaches (building on AFNI dmUBLOCK):
  1) "evtmax_1": pre-convolution HRF max=1.0 normalization of each stimulus regardless of duration: identical to dmUBLOCK(1)
  2) "durmax_1": pre-convolution HRF max=1.0 normalization for long events (15+ sec) -- height of HRF is modulated by duration of event: identical to dmUBLOCK(0)
}
\author{
Michael Hallquist
}
\keyword{internal}
