% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/job_tracking_functions.R
\name{create_tracking_db}
\alias{create_tracking_db}
\title{Internal helper function to create the tracking SQLite database}
\usage{
create_tracking_db(sqlite_db)
}
\arguments{
\item{sqlite_db}{Path to SQLite database used for tracking}
}
\description{
Internal helper function to create the tracking SQLite database
}
\keyword{internal}
