% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deconvolve_nlreg_resample.R
\name{deconvolve_nlreg_resample}
\alias{deconvolve_nlreg_resample}
\title{This function deconvolves the BOLD signal using Bush 2011 method, augmented
by the resampling approach of Bush 2015.}
\usage{
deconvolve_nlreg_resample(
  bold_obs,
  kernel,
  nev_lr = 0.01,
  epsilon = 0.005,
  beta = 40,
  n_resample = 30,
  trim_kernel = TRUE
)
}
\arguments{
\item{bold_obs}{observed BOLD signal}

\item{kernel}{assumed kernel of the BOLD signal}

\item{nev_lr}{learning rate for the assignment of neural events. Default: .01}

\item{epsilon}{relative error change (termination condition). Default: .005}

\item{beta}{slope of the sigmoid transfer function. Default: 40}

\item{n_resample}{number of resampling steps for deconvolution. Default: 30}

\item{trim_kernel}{whether to remove the first K time points from the deconvolved vector, corresponding to
kernel leftovers from convolution. Default: TRUE}
}
\value{
list containing the following fields:
    - NEVest  - the base neural event estimate
    - NEVmean - the mean neural event estimate
    - NEVstd - the std dev. of the neural event estimate
    - NEVcupp - the mean (upper limit 95% ci)
    - NEVclow - the mean (lower limit 95% ci)
    - BLDmean - the mean of the BOLD estimate
    - BLDstd - the std dev. of the BOLD estimate
    - BLDcupp - the mean BOLD (upper limit 95% ci)
    - BLDclow - the mean BOLD (lower limit 95% ci)
}
\description{
This function deconvolves the BOLD signal using Bush 2011 method, augmented
by the resampling approach of Bush 2015.
}
\details{
Author: Keith Bush
Institution: University of Arkansas at Little Rock
Date:        Aug. 12, 2013
}
\author{
Keith Bush
}
