% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmri_utility_fx.R
\name{detrendts}
\alias{detrendts}
\title{Detrend a time series up to quadratic trend. Used by fir1Bandpass prior to filtering}
\usage{
detrendts(x, order = 0)
}
\arguments{
\item{x}{A time series to be detrended}

\item{order}{The polynomial order used for detrending. 0=demean; 1=linear; 2=quadratic}
}
\description{
Detrend a time series up to quadratic trend. Used by fir1Bandpass prior to filtering
}
