% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_glm_pipeline.R
\name{dhms}
\alias{dhms}
\title{small helper function to parse duration syntax of days-hours:minutes:seconds
  into lubridate duration object}
\usage{
dhms(str)
}
\arguments{
\item{str}{string containing a duration that may include a days specification}
}
\description{
small helper function to parse duration syntax of days-hours:minutes:seconds
  into lubridate duration object
}
\keyword{internal}
