% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{do_convolve}
\alias{do_convolve}
\title{Internal port of fsl do_convolve}
\arguments{
\item{input}{the input}

\item{kernel}{the kernel to convolve}

\item{phase}{an integer for phase-shifting the HRF}

\item{renorm}{boolean indicating whether to renormalize the output by the sum}
}
\value{
A vector containing the convolution of input and kernel
}
\description{
Internal port of fsl do_convolve
}
\details{
This is an internal function used for testing alignment with FSL HRF convolution
}
