% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{dsigmoid}
\alias{dsigmoid}
\title{Dsigmoid transform}
\arguments{
\item{x}{value to be transformed}

\item{beta}{slope (steepness) of sigmoid transform}
}
\description{
Dsigmoid transform
}
\keyword{internal}
