% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/afni_3dlmer.R
\name{emmeans_to_3dlmer_glt}
\alias{emmeans_to_3dlmer_glt}
\title{Translator from emmeans-style contrast spec to 3dLMEr -gltCode}
\usage{
emmeans_to_3dlmer_glt(
  mobj,
  data,
  qVars = NULL,
  raw_glt_codes = NULL,
  context = NULL
)
}
\arguments{
\item{mobj}{high-level model specification object (hi_model_spec)}

\item{data}{the data.frame used for the model (to check factor levels)}
}
\value{
a list of 3dLMEr gltCode strings
}
\description{
Translator from emmeans-style contrast spec to 3dLMEr -gltCode
}
\keyword{internal}
