% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glm_helper_functions.R
\name{enforce_glms_complete}
\alias{enforce_glms_complete}
\title{Helper function to create named data.frame from a set of objects}
\usage{
enforce_glms_complete(gpa, level = 1L, lg = NULL)
}
\description{
Helper function to create named data.frame from a set of objects
}
\details{
This function helps with the problem of having several vectors
  in the workspace that you want to combine with data.fram
}
