% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/event_lock_timeseries.R
\name{event_lock_ts}
\alias{event_lock_ts}
\title{function to get interpolated event locked data}
\usage{
event_lock_ts(
  fmri_obj,
  event = NULL,
  time_before = -3,
  time_after = 3,
  collide_before = NULL,
  collide_after = NULL,
  pad_before = -1,
  pad_after = 1,
  time_audit = FALSE
)
}
\arguments{
\item{fmri_obj}{An fmri_ts object containing time series data that are optionally keyed on one or more
grouping variables (e.g., ROI). The fmri_obj must also contain trial-level data (in the $event_data field)
for event-locking to proceed.}

\item{event}{Name of column in \code{fmri_obj$event_data} that identifies onset for event of interest}

\item{time_before}{How many seconds before the \code{event} do we want data. Default: -3}

\item{time_after}{How many seconds after the \code{event} do we want data. Default: 3}

\item{collide_before}{An optional vector of column names in \code{trial_df} that set a boundary on the earliest
time point used in interpolation. This effectively truncates the time series for interpolation to a smaller window
than specified by \code{time_before}.}

\item{collide_after}{An optional vector of column names in \code{trial_df} that set a boundary on the latest
time point used in interpolation. This effectively truncates the time series for interpolation to a smaller window
than specified by \code{time_after}.}

\item{pad_before}{Number of seconds to include in the epoch time window before the event of interest. Interpolation
spans the window from \code{time_before} to \code{time_after}, but padding includes data points at the
boundary that can help to have sufficient data to interpolate early and late times within the epoch.}

\item{pad_after}{Number of seconds to include in the epoch time window after the event of interest.}

\item{time_audit}{If TRUE, additional columns will be added to the output showing how alignment is calculated vis-a-vis event timing}

\item{logfile}{Name of log file containing event-locking problems}
}
\description{
function to get interpolated event locked data
}
