% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yaml_outputs.R
\name{export_l1_config}
\alias{export_l1_config}
\title{Helper function to output l1_models into '.yaml' object}
\usage{
export_l1_config(gpa, file = "l1_config.yaml")
}
\arguments{
\item{gpa}{the \code{gpa} object}
}
\value{
a yaml document of l1_models for an object
}
\description{
Helper function to output l1_models into '.yaml' object
}
