% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_glm_betas_in_mask.R
\name{extract_fsl_betas}
\alias{extract_fsl_betas}
\title{function to extract voxelwise betas from a given level of FSL analysis}
\usage{
extract_fsl_betas(
  gpa,
  extract = NULL,
  level = NULL,
  what = c("cope", "zstat"),
  aggregate = TRUE,
  aggFUN = "mean",
  remove_zeros = TRUE,
  mask_file = NULL,
  ncores = 1L,
  lg = NULL
)
}
\arguments{
\item{gpa}{a \code{glm_pipeline_arguments} object}

\item{level}{which level to extract: 1, 2, 3}

\item{what}{which elements of the FEAT output should be extracted. Default is cope and zstat}

\item{aggregate}{whether to aggregate voxels within each parcel}

\item{aggFUN}{the function to use for aggregating voxels within a parcel}

\item{remove_zeros}{whether to remove statistics that are very close to zero (and may thus represent masked-out voxels)}

\item{mask_file}{a nifti image containing integer values for each parcel from which we should extract statistics}

\item{ncores}{the number of cores to use for image extraction. Default: 1}

\item{lg}{a Logger object for logging beta extraction}
}
\description{
function to extract voxelwise betas from a given level of FSL analysis
}
\keyword{internal}
