% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_glm_betas_atlas.R
\name{extract_glm_betas_atlas}
\alias{extract_glm_betas_atlas}
\title{Function to extract subject-level coefficients from an FSL group analysis}
\usage{
extract_glm_betas_atlas(
  config_file,
  atlas_files,
  run_model_index = 1,
  extract_z = FALSE,
  extract_beta_series = FALSE,
  ncpus = 1,
  aggregate = TRUE,
  aggFUN = mean
)
}
\arguments{
\item{atlas_files}{vector of filenames for NIfTI-based atlases. The function will loop over each and extract coefficients}

\item{extract_z}{extract lower-level z-statistics if TRUE, or copes (betas) if FALSE}

\item{extract_beta_series}{whether to extract coefficients from parallel beta series analyses that have already been run (very slow)}

\item{aggregate}{whether to take the average (or other central tendency measure) of voxels within a given mask value}
}
\description{
Function to extract subject-level coefficients from an FSL group analysis
}
