% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/populate_l1_from_yaml.R
\name{fields_from_spec}
\alias{fields_from_spec}
\title{internal function to populate l1_model_set onsets from spec specification}
\usage{
fields_from_spec(l1_model_set, slist, trial_data, field_names = NULL)
}
\arguments{
\item{l1_model_set}{An L1 model set.}

\item{slist}{Parsed signal/model specification list.}

\item{trial_data}{Trial-level data.}

\item{field_names}{Fields to populate from the specification.}
}
\description{
internal function to populate l1_model_set onsets from spec specification
}
