% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glm_helper_functions.R
\name{file_ext}
\alias{file_ext}
\title{internal function for getting a file extension that may include a compressed ending
this is adapted from neurobase::file_imgext, but extended for other file types}
\usage{
file_ext(file, withdot = TRUE)
}
\description{
internal function for getting a file extension that may include a compressed ending
this is adapted from neurobase::file_imgext, but extended for other file types
}
\keyword{internal}
