% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glm_helper_functions.R
\name{file_sans_ext}
\alias{file_sans_ext}
\title{internal function for returning a file (and path) without its extension
at present, it returns NA if no recognized extension is there}
\usage{
file_sans_ext(file, withdot = TRUE)
}
\description{
internal function for returning a file (and path) without its extension
at present, it returns NA if no recognized extension is there
}
\keyword{internal}
