% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fill_atlas_with_stats.R
\name{fill_atlas_with_stats}
\alias{fill_atlas_with_stats}
\title{Populates parcel-wise statistics from an atlas or set of clusters back into NIfTIs in the original space}
\usage{
fill_atlas_with_stats(
  atlas_nifti,
  stat_dt,
  stat_cols = c("t", "p"),
  stat_labels = NULL,
  atlas_col = "atlas_value",
  split_on = NULL,
  stack_along = NULL,
  img_prefix = "atlasfill",
  out_dir = getwd(),
  overwrite = FALSE,
  afni_dir = "~/abin"
)
}
\arguments{
\item{atlas_nifti}{The filename of the NIfTI image containing atlas values to match against \code{stat_dt}}

\item{stat_dt}{a \code{data.table} object containing the statistics to write to parcels in the atlas}

\item{stat_cols}{The column names in \code{stat_dt} that should be written to parcels in the NIfTI. By default,
each column will be written as a separate sub-brik (i.e., along the 4th dimension of the NIfTI), but this can be
modified by the \code{stack_along} argument.}

\item{stat_labels}{A character vector of the same length as \code{stat_cols} giving the names for each statistic
that will be used in the output NIfTI files. By default the labels are simply the \code{stat_cols}.}

\item{atlas_col}{The column name in \code{stat_dt} containing the integer values that correspond to parcel numbers
in \code{atlas_nifti}. This is used to map the statistics back into the \code{atlas_nifti} space.}

\item{split_on}{An optional character vector that yields an output image for each unique combination of splits. For
example, if \code{stat_dt} contains multiple GLM contrasts in \code{"contrast"} and multiple subjects in \code{"subject"},
then \code{split_on = c("contrast", "subject")} would yield a separate NIfTI with the statistics for each contrast and
subject. This allows the fill operation to be repeated over an arbitrarily stacked dataset.}

\item{stack_along}{By default, this is \code{NULL}, in which case each output NIfTI will have a sub-briks for each statistic
in \code{stat_cols}. Alternatively, \code{stack_along} can be a character vector of other columns in \code{stat_dt}. In this case,}

\item{img_prefix}{A string prefixing the filenames for all NIfTIs}

\item{out_dir}{The directory in which to place the stat NIfTIs. If not present, it will be created}

\item{overwrite}{Whether to allow existing files to be overwritten. Default: FALSE}

\item{afni_dir}{The location of your AFNI installation, which is used to cobble together NIfTIs}
}
\description{
Populates parcel-wise statistics from an atlas or set of clusters back into NIfTIs in the original space
}
