% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finalize_pipeline_configuration.R
\name{finalize_confound_settings}
\alias{finalize_confound_settings}
\title{Helper function to populate confound information for pipeline files}
\usage{
finalize_confound_settings(gpa, lg)
}
\arguments{
\item{gpa}{a glm_pipeline_arguments object for population}

\item{lg}{a Logger object for logging results of confound processing}
}
\description{
Helper function to populate confound information for pipeline files
}
\keyword{internal}
