% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finalize_pipeline_configuration.R
\name{finalize_pipeline_configuration}
\alias{finalize_pipeline_configuration}
\title{This is a small helper function to validate the glm_model_arguments list structure.
It adds a few details such as the output directory to make it less burdensome for to setup a pipeline
N.B. gpa is a shorthand abbreviation for glm_model_arguments, to save typing}
\usage{
finalize_pipeline_configuration(gpa)
}
\arguments{
\item{gpa}{A \code{glm_pipeline_arguments} object setup by \code{setup_glm_pipeline}}
}
\description{
This is a small helper function to validate the glm_model_arguments list structure.
It adds a few details such as the output directory to make it less burdensome for to setup a pipeline
N.B. gpa is a shorthand abbreviation for glm_model_arguments, to save typing
}
