% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmri_utility_fx.R
\name{fir1Bandpass}
\alias{fir1Bandpass}
\title{Apply a FIR-1 bandpass filter a signal. Can low- or high-pass filter by specifying 0 for low or >= Nyquist for high.}
\usage{
fir1Bandpass(
  x,
  TR = 2,
  low = 0.009,
  high = 0.08,
  n = 250,
  plotFilter = FALSE,
  forward_reverse = TRUE,
  padx = 0,
  detrend = 1
)
}
\arguments{
\item{x}{The time series to be filtered}

\item{TR}{The sampling frequency in seconds}

\item{low}{The lower filter cutoff in Hz. Fluctuations below this frequency will be filtered out}

\item{high}{The upper filter cutoff in Hz. Fluctuations above this frequency will be filtered out}

\item{n}{The order of the filter coefficients. Should probably leave this alone in general}
}
\description{
Apply a FIR-1 bandpass filter a signal. Can low- or high-pass filter by specifying 0 for low or >= Nyquist for high.
}
\keyword{internal}
