% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmri_utility_fx.R
\name{fmri.design}
\alias{fmri.design}
\title{fmri.design function from \code{fmri} R package. Brought into this package to avoid dependency on \code{fmri} package,
  especially the \code{gsl} package dependency in \code{aws}, which is finicky on clusters if the GSL module isn't available.}
\usage{
fmri.design(stimulus, order = 2, cef = NULL, verbose = FALSE)
}
\arguments{
\item{stimulus}{The regressor of interest}

\item{order}{The polynomial order of baseline/drift terms to incorporate in the design matrix}

\item{cef}{confound regressors to add}

\item{verbose}{if TRUE, output detailed progress about each step in the design setup}
}
\description{
fmri.design function from \code{fmri} R package. Brought into this package to avoid dependency on \code{fmri} package,
  especially the \code{gsl} package dependency in \code{aws}, which is finicky on clusters if the GSL module isn't available.
}
\keyword{internal}
