% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmri_utility_fx.R
\name{fmri.stimulus}
\alias{fmri.stimulus}
\title{Convolve a regressor with a hemodynamic response function for fMRI analysis.}
\usage{
fmri.stimulus(
  n_vols = 1,
  onsets = c(1),
  durations = c(1),
  values = c(1),
  times = NULL,
  center_values = FALSE,
  rm_zeros = TRUE,
  convolve = TRUE,
  tr = 2,
  ts_multiplier = NULL,
  demean = TRUE,
  convmax_1 = FALSE,
  a1 = 6,
  a2 = 12,
  b1 = 0.9,
  b2 = 0.9,
  cc = 0.35,
  conv_method = "r",
  microtime_resolution = 20
)
}
\arguments{
\item{n_vols}{The number of volumes (scans) to be output in the convolved regressor}

\item{onsets}{A vector of times (in scans) specifying event onsets}

\item{durations}{A vector of durations (in seconds) for each event}

\item{values}{A vector of parametric values used as regressor heights prior to convolution}

\item{times}{A vector of times (in seconds) specifying event onsets. If times are passed in, the onsets
argument (which is in scans) is ignored. That is, \code{onsets} and \code{times} are mutually exclusive.}

\item{center_values}{Whether to demean values vector before convolution}

\item{rm_zeros}{Whether to remove zeros from events vector prior to convolution. Generally a good idea since we typically center
values prior to convolution, and retaining zeros will lead them to be non-zero after mean centering.}

\item{convolve}{Whether to convolve the regressor with the HRF. If FALSE, a time series of events, durations, and heights is returned.}

\item{tr}{The repetition time (sometimes called TR) in seconds}

\item{ts_multiplier}{A time series of length \code{n_vols} that is multiplied against the stimulus before convolution.}

\item{demean}{Whether to demean the regressor after convolution. Default: TRUE}

\item{convmax_1}{Whether to rescale the convolved regressor to a maximum height of 1.}

\item{a1}{The a1 parameter of the double gamma}

\item{a2}{The a2 parameter of the double gamma}

\item{b1}{The b1 parameter of the double gamma}

\item{b2}{The b2 parameter of the double gamma}

\item{cc}{The cc parameter of the double gamma}

\item{conv_method}{Method for convolving HRF with stimulus. Either "r" or "cpp". The "r" method uses an FFT-based
internal convolution with convolve(x, y, conj=TRUE). The "cpp" method uses an internal C++ function with a
loop-based convolution over the vectors. Unfortunately, at present, the C++ approach is noticeably slower since
it does not use FFT to obtain the filter.}

\item{microtime_resolution}{The number of bins between TRs used for calculating regressor values in continuous time}
}
\description{
Extended from \code{fmri} package to allow for continuous-valued regressor,
which is passed using the values parameter.
}
\details{
The function also supports mean centering of parametric regressor prior to convolution
to dissociate it from stimulus occurrence (when event regressor also in model)
}
