% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fsl_l1_model.R
\name{fsl_l1_model}
\alias{fsl_l1_model}
\title{This function specifies a first-level GLM model in FSL based on a build_design_matrix
  bdm object.}
\usage{
fsl_l1_model(
  id = NULL,
  session = NULL,
  l1_confound_files = NULL,
  d_obj,
  gpa,
  model_name = NULL,
  run_nifti = NULL,
  nvoxels = NULL,
  execute_feat = FALSE
)
}
\arguments{
\item{d_obj}{a single-subject design matrix object generated by build_design_matrix}

\item{gpa}{a gpa (glm_pipeline_arguments) object generated by setup_glm_pipeline}

\item{model_name}{a string indicating the model name within \code{gpa} to setup. If
you wish to setup multiple models, this is handled upstream in setup_l1_models.R}

\item{run_nifti}{an optional character vector of NIfTI filenames used in l1 analysis.
Prefer to pass these via the build_design_matrix object in $run_nifti}
}
\description{
This function specifies a first-level GLM model in FSL based on a build_design_matrix
  bdm object.
}
\author{
Michael Hallquist
}
