% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fsl_l3_model.R
\name{fsl_l3_model}
\alias{fsl_l3_model}
\title{Estimate a level 2 (subject) model using FSL FEAT with fixed effects integration of runs}
\usage{
fsl_l3_model(l3_df = NULL, gpa)
}
\arguments{
\item{l3_df}{a data.frame containing cope inputs for a given l3 model, as well as metadata
that identify the model.}

\item{gpa}{a \code{glm_pipeline_arguments} object containing model specification}
}
\description{
Estimate a level 2 (subject) model using FSL FEAT with fixed effects integration of runs
}
\author{
Michael Hallquist
}
