% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/afni_clustsim.R
\name{fwhmx}
\alias{fwhmx}
\title{R6 class for running 3dFWHMx on a single input file based on user specification}
\description{
R6 class for running 3dFWHMx on a single input file based on user specification

R6 class for running 3dFWHMx on a single input file based on user specification
}
\details{
N.B. This class doesn't even expose the Gaussian ACF options given false positive problems
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{fwhmx$new()}}
\item \href{#method-run}{\code{fwhmx$run()}}
\item \href{#method-get_call}{\code{fwhmx$get_call()}}
\item \href{#method-get_acf_params}{\code{fwhmx$get_acf_params()}}
\item \href{#method-get_input_file}{\code{fwhmx$get_input_file()}}
\item \href{#method-get_mask_file}{\code{fwhmx$get_mask_file()}}
\item \href{#method-get_outputs}{\code{fwhmx$get_outputs()}}
\item \href{#method-is_fwhmx_complete}{\code{fwhmx$is_fwhmx_complete()}}
\item \href{#method-clone}{\code{fwhmx$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{fwhmx$new(
  input_file = NULL,
  mask_file = NULL,
  out_dir = NULL,
  demed = NULL,
  unif = NULL,
  average = "geometric"
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-run"></a>}}
\if{latex}{\out{\hypertarget{method-run}{}}}
\subsection{Method \code{run()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{fwhmx$run(force = FALSE)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_call"></a>}}
\if{latex}{\out{\hypertarget{method-get_call}{}}}
\subsection{Method \code{get_call()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{fwhmx$get_call()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_acf_params"></a>}}
\if{latex}{\out{\hypertarget{method-get_acf_params}{}}}
\subsection{Method \code{get_acf_params()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{fwhmx$get_acf_params()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_input_file"></a>}}
\if{latex}{\out{\hypertarget{method-get_input_file}{}}}
\subsection{Method \code{get_input_file()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{fwhmx$get_input_file()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_mask_file"></a>}}
\if{latex}{\out{\hypertarget{method-get_mask_file}{}}}
\subsection{Method \code{get_mask_file()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{fwhmx$get_mask_file()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_outputs"></a>}}
\if{latex}{\out{\hypertarget{method-get_outputs}{}}}
\subsection{Method \code{get_outputs()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{fwhmx$get_outputs()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-is_fwhmx_complete"></a>}}
\if{latex}{\out{\hypertarget{method-is_fwhmx_complete}{}}}
\subsection{Method \code{is_fwhmx_complete()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{fwhmx$is_fwhmx_complete()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{fwhmx$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
