% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/afni_clustsim.R
\name{fwhmx_set_spec}
\alias{fwhmx_set_spec}
\title{R6 class for running 3dFWHMx on a group of input files using a scheduler/cluster}
\description{
R6 class for running 3dFWHMx on a group of input files using a scheduler/cluster

R6 class for running 3dFWHMx on a group of input files using a scheduler/cluster
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{fwhmx_set_spec$new()}}
\item \href{#method-refresh}{\code{fwhmx_set_spec$refresh()}}
\item \href{#method-run}{\code{fwhmx_set_spec$run()}}
\item \href{#method-get_batch}{\code{fwhmx_set_spec$get_batch()}}
\item \href{#method-get_acf_average}{\code{fwhmx_set_spec$get_acf_average()}}
\item \href{#method-get_acf_df}{\code{fwhmx_set_spec$get_acf_df()}}
\item \href{#method-get_effective_fwhm}{\code{fwhmx_set_spec$get_effective_fwhm()}}
\item \href{#method-is_fwhmx_complete}{\code{fwhmx_set_spec$is_fwhmx_complete()}}
\item \href{#method-clone}{\code{fwhmx_set_spec$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{fwhmx_set_spec$new(
  input_files = NULL,
  mask_files = NULL,
  scheduler = NULL,
  walltime = "10:00:00",
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-refresh"></a>}}
\if{latex}{\out{\hypertarget{method-refresh}{}}}
\subsection{Method \code{refresh()}}{
recreate fwhmx objects and completion status
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{fwhmx_set_spec$refresh(...)}\if{html}{\out{</div>}}
}

\subsection{Details}{
useful for updating job status and ACF params after a run completes
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-run"></a>}}
\if{latex}{\out{\hypertarget{method-run}{}}}
\subsection{Method \code{run()}}{
Run the 3dFWHMx batch for all inputs
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{fwhmx_set_spec$run(force = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{force}}{If \code{TRUE}, completed 3dFWHMx runs will be included in the batch. Default: FALSE}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_batch"></a>}}
\if{latex}{\out{\hypertarget{method-get_batch}{}}}
\subsection{Method \code{get_batch()}}{
Return R_batch_job objects used to run 3dFWHMx for all inputs
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{fwhmx_set_spec$get_batch(include_complete = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{include_complete}}{If TRUE, complete 3dFWHMx jobs will be included in the batch, leading these to be re-run}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_acf_average"></a>}}
\if{latex}{\out{\hypertarget{method-get_acf_average}{}}}
\subsection{Method \code{get_acf_average()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{fwhmx_set_spec$get_acf_average(allow_incomplete_fwhmx = FALSE)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_acf_df"></a>}}
\if{latex}{\out{\hypertarget{method-get_acf_df}{}}}
\subsection{Method \code{get_acf_df()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{fwhmx_set_spec$get_acf_df()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_effective_fwhm"></a>}}
\if{latex}{\out{\hypertarget{method-get_effective_fwhm}{}}}
\subsection{Method \code{get_effective_fwhm()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{fwhmx_set_spec$get_effective_fwhm()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-is_fwhmx_complete"></a>}}
\if{latex}{\out{\hypertarget{method-is_fwhmx_complete}{}}}
\subsection{Method \code{is_fwhmx_complete()}}{
Simple method to return whether 3dFWHMx is complete for all input files
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{fwhmx_set_spec$is_fwhmx_complete()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{fwhmx_set_spec$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
