% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{generate_feature_armadillo}
\alias{generate_feature_armadillo}
\title{This function creates K shifts of a neural events vector according to the kernel length, K.}
\arguments{
\item{encoding}{The neural events vector (same length as BOLD time series)}

\item{K}{The length of the kernel}
}
\value{
A matrix of length(encoding) rows and K columns, where each column contains a successively
   lagged copy of the encoding vector
}
\description{
This function creates K shifts of a neural events vector according to the kernel length, K.
}
\details{
This is an internal function that is used inside a while loop by deconvolve_nlreg.
  Profiling of the algorithm revealed that this is the primary bottleneck, so I ported it to
  an Rcpp function
}
\author{
Michael Hallquist
}
