% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spm_contrasts.R
\name{generate_spm_contrasts_from_model}
\alias{generate_spm_contrasts_from_model}
\title{Generate SPM contrasts from an fmri.pipeline L1 model contrast matrix}
\usage{
generate_spm_contrasts_from_model(
  output_dir,
  mobj,
  spm_path = "/gpfs/group/mnh5174/default/lab_resources/spm12",
  execute = FALSE,
  matlab_cmd = "matlab",
  matlab_args = "-batch",
  average_across_runs = TRUE
)
}
\arguments{
\item{output_dir}{location for SPM outputs and scripts for estimating contrasts. Must contain SPM.mat already}

\item{mobj}{l1_model_spec object containing $contrasts and regressor names}

\item{spm_path}{see generate_spm_mat}

\item{execute}{whether to run contrast setup. This depends on SPM.mat having been created already. Default: FALSE}

\item{matlab_cmd}{see generate_spm_mat}

\item{matlab_args}{see generate_spm_mat}

\item{average_across_runs}{whether to average regressor weights across run-specific columns in SPM.xX}
}
\description{
Generate SPM contrasts from an fmri.pipeline L1 model contrast matrix
}
