% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bids_tools.R
\name{generate_subject_data_from_bids}
\alias{generate_subject_data_from_bids}
\title{Function to generate a subject_data object from a BIDS-compliant folder}
\usage{
generate_subject_data_from_bids(bids_dir)
}
\arguments{
\item{bids_dir}{a directory containing BIDS-compliant processed data for analysis}
}
\value{
a data.frame containing subject-level data contained in participants.tsv
}
\description{
Function to generate a subject_data object from a BIDS-compliant folder
}
\details{
This function basically just reads participants.tsv from the root of the BIDS folder and renames
  `participant_id`` to `id` to match this pipeline's conventions
}
