% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bids_tools.R
\name{generate_trial_data_from_bids}
\alias{generate_trial_data_from_bids}
\title{Function to generate a trial_data object from a BIDS-compliant folder}
\usage{
generate_trial_data_from_bids(bids_dir, modality = "func", task_name = "ridl")
}
\arguments{
\item{bids_dir}{a directory containing BIDS-compliant processed data for analysis}

\item{modality}{the subfolder within \code{bids_dir} that contains data of a certain modality.
Almost always 'func', which is the default.}

\item{task_name}{the name of the task, which is matched as a BIDS \code{task-} entity.}
}
\value{
a data.frame containing all run_nifti and confound_input_file results for subjects in the folder
}
\description{
Function to generate a trial_data object from a BIDS-compliant folder
}
\details{
The files should generally have a name like
  sub-220256_task-ridl3_space-MNI152NLin2009cAsym_desc-preproc_bold_postprocessed.nii.gz
  and be located in a folder like: /proj/mnhallqlab/proc_data/sub-220256/func/
  where 'func' is the \code{modality}, and 'ridl' is the \code{task_name}.
  '_postprocessed' is the \code{suffix}.
}
\examples{
\dontrun{
  bids_dir <- "/proj/mnhallqlab/no_backup/flanker-fmriprep"
  df <- generate_trial_data_from_bids(
    bids_dir,
    task_name = "flanker"
  )
}
}
