% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/afni_3dlmer.R
\name{get_3dlmer_status}
\alias{get_3dlmer_status}
\title{Status checker for 3dLMEr}
\usage{
get_3dlmer_status(output_file, lg = NULL, prefix = NULL)
}
\arguments{
\item{output_file}{expected output file from 3dLMEr}

\item{lg}{optional logger}

\item{prefix}{optional prefix for returned column names}
}
\value{
data.frame indicating output existence and completion status
}
\description{
Status checker for 3dLMEr
}
\keyword{internal}
