% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_batch_job.R
\name{get_batch_job_status}
\alias{get_batch_job_status}
\title{query job status}
\usage{
get_batch_job_status(jobid = NULL, return_children = FALSE, sqlite_db)
}
\arguments{
\item{jobid}{the batch job id}

\item{sqlite_db}{character string of sqlite database}
}
\description{
query job status
}
