% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmri_utility_fx.R
\name{get_collin_events}
\alias{get_collin_events}
\title{Small helper function to calculate collinearity diagnostics on the events in the design
  prior to convolution. This can help to diagnose problems with parametric modulators that
  correlate excessively, for example.}
\usage{
get_collin_events(dmat)
}
\arguments{
\item{dmat}{A design matrix list generated by \code{build_design_matrix} that contains
a set of regressors for each run, where run is indicated by the rows of \code{dmat}}
}
\description{
Small helper function to calculate collinearity diagnostics on the events in the design
  prior to convolution. This can help to diagnose problems with parametric modulators that
  correlate excessively, for example.
}
\keyword{internal}
