% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glm_helper_functions.R
\name{get_confound_txt}
\alias{get_confound_txt}
\title{helper function to generate confounds txt file for inclusion as additional regressors}
\usage{
get_confound_txt(mr_file, gpa, drop_volumes = 0L, last_volume = NULL)
}
\arguments{
\item{mr_file}{an absolute path to a nifti file for a single run}

\item{gpa}{a \code{glm_pipeline_arguments} object containing pipeline specification}
}
\description{
helper function to generate confounds txt file for inclusion as additional regressors
}
