% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glm_helper_functions.R
\name{get_contrasts_from_spec}
\alias{get_contrasts_from_spec}
\title{helper function to generate a contrast matrix from an lm() object
  and a set of user-specified contrasts using emmeans}
\usage{
get_contrasts_from_spec(mobj, lmfit = NULL)
}
\arguments{
\item{mobj}{a model object created by build_l<X>_models}

\item{lmfit}{an optional lm() object used for emmeans calculations. If provided, this object
will be used instead of mobj$lmfit (the parent lm on the overall dataset).}
}
\value{
a modified copy of the model object \code{mobj} with $contrast_list and $contrasts
  fully populated
}
\description{
helper function to generate a contrast matrix from an lm() object
  and a set of user-specified contrasts using emmeans
}
\keyword{internal}
