% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sqlite_functions.R
\name{get_default_sqlite_busy_timeout}
\alias{get_default_sqlite_busy_timeout}
\title{helper function to get generic busy timeout}
\usage{
get_default_sqlite_busy_timeout(ms = T)
}
\arguments{
\item{ms}{Logical. Return in milliseconds? If FALSE, returns in seconds}
}
\description{
helper function to get generic busy timeout
}
\keyword{internal}
