% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fsl_helper_functions.R
\name{get_feat_dir_files}
\alias{get_feat_dir_files}
\title{internal helper function to look up core stats outputs from a .feat folder}
\usage{
get_feat_dir_files(feat_dir)
}
\arguments{
\item{feat_dir}{a .feat folder containing the outputs of an FSL analysis}
}
\value{
a list containing sorted vectors of each stat output
}
\description{
internal helper function to look up core stats outputs from a .feat folder
}
\keyword{internal}
