% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fsl_helper_functions.R
\name{get_feat_status}
\alias{get_feat_status}
\title{helper function to look at whether feat ingredients exist for a .feat/.gfeat directory}
\usage{
get_feat_status(feat_dir, feat_fsf = NULL, lg = NULL, prefix = NULL)
}
\arguments{
\item{feat_dir}{an expected .feat/.gfeat directory for an analysis}

\item{feat_fsf}{an expected .fsf file corresponding to the feat analysis}

\item{lg}{an optional lgr logger object to be used for logging. If not passed, the root
logger will be used}
}
\value{
a data.frame containing information about whether the .feat analysis is complete
  and whether various ingredients are present
}
\description{
helper function to look at whether feat ingredients exist for a .feat/.gfeat directory
}
\keyword{internal}
