% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wait_for_job.R
\name{get_job_status}
\alias{get_job_status}
\title{Get the status of a list of jobs based on scheduler type}
\usage{
get_job_status(job_ids = NULL, scheduler = "local")
}
\description{
Get the status of a list of jobs based on scheduler type
}
\author{
Michael Hallquist
}
