% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yaml_outputs.R
\name{get_l2_config}
\alias{get_l2_config}
\title{Helper function to output l2_models into '.yaml' object}
\usage{
get_l2_config(gpa)
}
\arguments{
\item{gpa}{the \code{gpa} object}
}
\value{
a yaml document of l2_models for an object
}
\description{
Helper function to output l2_models into '.yaml' object
}
\keyword{internal}
