% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yaml_outputs.R
\name{get_l3_config}
\alias{get_l3_config}
\title{Get a list containing configuration settings for l3 models}
\usage{
get_l3_config(gpa)
}
\description{
Get a list containing configuration settings for l3 models
}
\keyword{internal}
