% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/event_lock_timeseries.R
\name{get_medusa_compression_score}
\alias{get_medusa_compression_score}
\title{front-end function for taking a list of windowed time series by mask value, interpolating them onto a time grid,
and (optionally) averaging across voxels/units within a value to derive the mean interpolated time series}
\usage{
get_medusa_compression_score(
  fmri_obj,
  event = NULL,
  time_before = -3,
  time_after = 3,
  collide_before = NULL,
  collide_after = NULL,
  group_by = NULL
)
}
\arguments{
\item{fmri_obj}{an fmri_ts object containing a single run of data with corresponding events}

\item{event}{the event to which the fmri time series should be aligned (column in \code{fmri_obj$event_data})}

\item{time_before}{How many seconds before the \code{event} do we want data. Default: -3}

\item{time_after}{How many seconds after the \code{event} do we want data. Default: 3}

\item{collide_before}{An optional vector of column names in \code{trial_df} that set a boundary on the earliest
time point used in interpolation. This effectively truncates the time series for interpolation to a smaller window
than specified by \code{time_before}.}

\item{collide_after}{An optional vector of column names in \code{trial_df} that set a boundary on the latest
time point used in interpolation. This effectively truncates the time series for interpolation to a smaller window
than specified by \code{time_after}.}

\item{group_by}{return interpolated time series for each combination of group_by variables. Default is
to provide one interpolated time series per trial.}
}
\description{
front-end function for taking a list of windowed time series by mask value, interpolating them onto a time grid,
and (optionally) averaging across voxels/units within a value to derive the mean interpolated time series
}
\author{
Michael Hallquist
}
