% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_config.R
\name{get_sequence_info}
\alias{get_sequence_info}
\title{helper to retrieve information about a specific sequence from the config file}
\usage{
get_sequence_info(config_file = NULL, sequence_id = NULL)
}
\arguments{
\item{config_file}{the path to the configuration file}

\item{sequence_id}{unique identifier of sequence}
}
\description{
helper to retrieve information about a specific sequence from the config file
}
\keyword{internal}
