% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_sqlite_conn.R
\name{get_sqlite_conn}
\alias{get_sqlite_conn}
\title{Creates a SQLite database connection}
\usage{
get_sqlite_conn(db_path, sqlite_con = NULL)
}
\arguments{
\item{db_path}{Path to SQLite database}

\item{sqlite_con}{May pass existing DB connection to check}
}
\value{
DB connection
}
\description{
Creates a SQLite database connection
}
