% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnose_pipeline.R
\name{get_status_symbol}
\alias{get_status_symbol}
\title{helper for printing symbols based on status}
\usage{
get_status_symbol(status)
}
\arguments{
\item{status}{Character string job status}
}
\description{
helper for printing symbols based on status
}
\keyword{internal}
