% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnose_pipeline.R
\name{get_step_title}
\alias{get_step_title}
\title{helper for printing a more formal title for the steps in the pipeline}
\usage{
get_step_title(step)
}
\arguments{
\item{step}{Character string step name (as defined in run_glm_pipeline), e.g., 'run_l1'}
}
\description{
helper for printing a more formal title for the steps in the pipeline
}
\keyword{internal}
