% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logging_helpers.R
\name{get_subject_logger}
\alias{get_subject_logger}
\title{Create or retrieve a subject level logger}
\usage{
get_subject_logger(base_logger, id, gpa, log_prefix = "setup_l1_models")
}
\arguments{
\item{base_logger}{Name of parent logger (e.g., "glm_pipeline/l1_setup")}

\item{id}{Subject identifier}

\item{gpa}{glm_pipeline_arguments object containing logging settings}

\item{log_prefix}{Optional filename prefix for subject log files}
}
\description{
Create or retrieve a subject level logger
}
\keyword{internal}
