% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/job_tracking_functions.R
\name{get_top_parent}
\alias{get_top_parent}
\title{Helper to trace back to highest level parent from a given reference id}
\usage{
get_top_parent(tracking_df, ref_id)
}
\arguments{
\item{tracking_df}{a data.frame object extracted from the job tracking SQLite database}

\item{ref_id}{the id of the job from which to trace back}
}
\description{
Helper to trace back to highest level parent from a given reference id
}
\keyword{internal}
