% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/job_tracking_functions.R
\name{get_tracked_job_status}
\alias{get_tracked_job_status}
\title{Query job status in tracking SQLite database}
\usage{
get_tracked_job_status(
  job_id = NULL,
  return_children = FALSE,
  return_parent = FALSE,
  sequence_id = NULL,
  sqlite_db
)
}
\arguments{
\item{job_id}{The job id for which to retreive the status}

\item{return_children}{Return child jobs of this job}

\item{return_parent}{Return parent jobs of this job}

\item{sequence_id}{Alternatively return all jobs with sequence ID; cannot specify both job ID and sequence ID}

\item{sqlite_db}{Character string of sqlite database}
}
\value{
An R data.frame version of the tracking database
}
\description{
Query job status in tracking SQLite database
}
