% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggbrain.R
\name{ggbrain}
\alias{ggbrain}
\title{Plot fMRI data on an underlay image}
\usage{
ggbrain(
  underlay = NULL,
  overlay = NULL,
  color_col = NULL,
  alpha_col = NULL,
  underlay_colorscale = scale_fill_gradient(low = "grey10", high = "grey90"),
  overlay_colorscale = scale_fill_gradient2(midpoint = 0, low = "blue", mid = "grey90",
    high = "red"),
  coronal_slices = list(quantiles = c(0.25, 0.5, 0.75), ijk = NULL, xyz = NULL),
  sagittal_slices = list(quantiles = c(0.25, 0.5, 0.75), ijk = NULL, xyz = NULL),
  axial_slices = list(quantiles = c(0.25, 0.5, 0.75), ijk = NULL, xyz = NULL),
  remove_null_space = TRUE,
  pos_thresh = 3,
  neg_thresh = -3,
  legend_label = "z",
  background_color = "gray10",
  text_color = "white",
  trim_underlay = c(0.01, 0.99),
  zero_underlay = 0.001,
  symmetric_legend = TRUE,
  add_panel_labels = FALSE
)
}
\arguments{
\item{underlay}{a 3D nifti image used for the image underlay (default b/w)}

\item{overlay}{a 4D nifti image used for plotting stats on underlay (color)}

\item{color_col}{a position in the 4th dim of overlay use to color plots}

\item{alpha_col}{a position in the 4th dim of overlay use to set alpha transparency of plots}

\item{underlay_colorscale}{A ggplot scale_fill_* function call used for coloration of underlay}

\item{overlay_colorscale}{A ggplot scale_fill_* function call used for coloration of overlay}

\item{coronal_slices}{A list of coronal slices to be displayed. Elements can be \code{xyz}, \code{ijk},
and/or \code{quantiles}.}

\item{sagittal_slices}{A list of sagittal slices to be displayed. Elements can be \code{xyz}, \code{ijk},
and/or \code{quantiles}.}

\item{axial_slices}{A list of axial slices to be displayed. Elements can be \code{xyz}, \code{ijk},
and/or \code{quantiles}}

\item{remove_null_space}{If TRUE, quantiles are computed on the non-zero slices and plots are trimmed for
empty space.}

\item{pos_thresh}{The positive threshold to be applied to the \code{overlay} image. Any voxel that exceeds
the}

\item{trim_underlay}{Winsorize the extreme values of the underlay based on the low and high quantiles provided}

\item{zero_underlay}{Any voxels in the underlay image whose absolute values are less than \code{zero_underlay}
are set to precisely zero. This helps with small color variation in black/empty space.}
}
\description{
Plot fMRI data on an underlay image
}
