% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glm_helper_functions.R
\name{hours_to_dhms}
\alias{hours_to_dhms}
\title{convert a number of hours to a days, hours, minutes, seconds format}
\usage{
hours_to_dhms(hours, frac = FALSE)
}
\description{
convert a number of hours to a days, hours, minutes, seconds format
}
\keyword{internal}
