% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insert_df_sqlite.R
\name{insert_df_sqlite}
\alias{insert_df_sqlite}
\title{helper function to insert a keyed data.frame into the sqlite storage database}
\usage{
insert_df_sqlite(
  gpa = NULL,
  id = NULL,
  session = NULL,
  run_number = NULL,
  data = NULL,
  table = NULL,
  delete_extant = TRUE,
  append = TRUE,
  overwrite = FALSE
)
}
\arguments{
\item{id}{the id of the subject to whom these data belong}

\item{run_number}{the run_number of these data}

\item{data}{A \code{data.frame} containing the data to be inserted into the sqlite db}

\item{table}{A character string of the table name to be modified}

\item{delete_extant}{Whether to delete any existing records for this id + session + run_number combination}

\item{append}{Whether to append records to the table (passed through to dbWriteTable)}

\item{overwrite}{Whether to overwrite the existing table (passed through to dbWriteTable)}

\item{sesssion}{the session of these data}
}
\value{
a TRUE/FALSE indicating whether the record was successfully inserted
}
\description{
helper function to insert a keyed data.frame into the sqlite storage database
}
