% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/job_tracking_functions.R
\name{insert_tracked_job}
\alias{insert_tracked_job}
\title{Internal helper funciton to insert a job into the tracking SQLite database}
\usage{
insert_tracked_job(sqlite_db, job_id, tracking_args = list())
}
\arguments{
\item{sqlite_db}{Path to SQLite database used for tracking}
}
\description{
Internal helper funciton to insert a job into the tracking SQLite database
}
\keyword{internal}
