% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/event_lock_timeseries.R
\name{interpolate_fmri_epochs}
\alias{interpolate_fmri_epochs}
\title{Interpolate an aligned fmri_ts object onto a consistent time grid with respect to a target event}
\usage{
interpolate_fmri_epochs(
  a_obj,
  evt_time = "evt_time",
  time_before = -3,
  time_after = 3,
  output_resolution = 1,
  group_by = NULL
)
}
\arguments{
\item{a_obj}{an fmri_ts object that has data aligned to an event of interest}

\item{evt_time}{column name in \code{a_obj} event data to which time series should be aligned and interpolated}

\item{time_before}{The earliest time point (in seconds) relative to the event to be output in interpolation}

\item{time_after}{The latest time point (in seconds) relative to the event to be output in interpolation}

\item{output_resolution}{The timestep (in seconds) used for interpolation}

\item{group_by}{a character vector of keying variables used for aggregation of data prior to interpolation}
}
\description{
Interpolate an aligned fmri_ts object onto a consistent time grid with respect to a target event
}
\author{
Michael Hallquist
}
\keyword{internal}
