% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fsl_l2_model.R
\name{is_l2_passthrough_estimable}
\alias{is_l2_passthrough_estimable}
\title{Estimate a level 2 (subject) model using FSL FEAT with fixed effects integration of runs}
\usage{
is_l2_passthrough_estimable(dmat, cmat, tol = sqrt(.Machine$double.eps))
}
\arguments{
\item{dmat}{L2 design matrix.}

\item{cmat}{L2 contrast matrix.}

\item{tol}{Numerical tolerance for estimability checks.}
}
\description{
Estimate a level 2 (subject) model using FSL FEAT with fixed effects integration of runs
}
\author{
Michael Hallquist
}
\keyword{internal}
