% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yaml_outputs.R
\name{l1_yaml}
\alias{l1_yaml}
\title{Helper function to output l1_models into '.yaml' object}
\usage{
l1_yaml(gpa)
}
\arguments{
\item{gpa}{the \code{gpa} object}
}
\value{
a yaml document of l1_models for an object
}
\description{
Helper function to output l1_models into '.yaml' object
}
