% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glm_helper_functions.R
\name{lookup_dim}
\alias{lookup_dim}
\title{Helper function to obtain the dimensions of a NIfTI file}
\usage{
lookup_dim(nifti)
}
\arguments{
\item{nifti}{a nifti file}
}
\value{
a named vector of dimensions for the input (dim_x, dim_y, dim_z, dim_t)
}
\description{
Helper function to obtain the dimensions of a NIfTI file
}
\keyword{internal}
