% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookup_nifti_inputs.R
\name{lookup_nifti_inputs}
\alias{lookup_nifti_inputs}
\title{helper function to locate nifti inputs for each run of data}
\usage{
lookup_nifti_inputs(gpa, add_run_volumes = TRUE)
}
\arguments{
\item{gpa}{a \code{glm_pipeline_arguments} object created by setup_glm_pipeline}
}
\value{
a modified version of the \code{gpa} object that contains a $run_nifti field in
  the $run_data data.frame.
}
\description{
helper function to locate nifti inputs for each run of data
}
\author{
Michael Hallquist
}
\keyword{internal}
