% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glm_helper_functions.R
\name{lookup_nvoxels}
\alias{lookup_nvoxels}
\title{Helper function to obtain the number of voxels in a 4D file for populating FEAT FSF}
\usage{
lookup_nvoxels(nifti)
}
\arguments{
\item{nifti}{a 4D nifti file}
}
\value{
the number of voxels in \code{nifti} as calculated by x * y * z * t
}
\description{
Helper function to obtain the number of voxels in a 4D file for populating FEAT FSF
}
\keyword{internal}
