% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glm_helper_functions.R
\name{lookup_run_volumes}
\alias{lookup_run_volumes}
\title{Helper function to obtain the number of volumes in a 4D nifti file}
\usage{
lookup_run_volumes(nifti)
}
\arguments{
\item{nifti}{a 4D nifti file}
}
\value{
the number of volumes in \code{nifti}
}
\description{
Helper function to obtain the number of volumes in a 4D nifti file
}
\details{
This function prefers to use fslval instead of an internal R library
  because both oro.nifti and RNifti are rather slow to obtain a single value from
  the NIfTI header
}
\keyword{internal}
