% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggbrain.R
\name{lookup_slices}
\alias{lookup_slices}
\title{internal function to lookup which slices to display along each axis based on their quantile,
  xyz coordinate, or ijk coordinate}
\usage{
lookup_slices(
  sagittal_slices = NULL,
  coronal_slices = NULL,
  axial_slices = NULL,
  underlay,
  remove_null_space = TRUE
)
}
\arguments{
\item{sagittal_slices}{List of sagittal slices to display}

\item{coronal_slices}{List of coronal slices to display}

\item{axial_slices}{List of axial slices to display}

\item{underlay}{the RNifti object of the underlay that defines the coordinate system}
}
\description{
internal function to lookup which slices to display along each axis based on their quantile,
  xyz coordinate, or ijk coordinate
}
\keyword{internal}
