% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{mat2df}
\alias{mat2df}
\title{Fast conversion of a 2D numeric matrix to a 3-column data.frame}
\usage{
mat2df(
  mat,
  na_zeros = FALSE,
  varnames = as.character(c("dim1", "dim2")),
  value_name = "value"
)
}
\arguments{
\item{mat}{A \code{matrix} to convert to data.frame}

\item{na_zeros}{Whether near-zero values should be converted to \code{NA}}

\item{varnames}{Names for the two dimension key columns}

\item{value_name}{Name for the value column}
}
\value{
A 3-column data.frame with two dimension key columns and one value column
}
\description{
Converts a 2D numeric matrix into a 3-column data.frame
}
\details{
This function is a fast matrix-to-long-data-frame converter for the simple 2D case.
}
\author{
Michael Hallquist
}
\keyword{internal}
