% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glm_helper_functions.R
\name{mobj_fit_lm}
\alias{mobj_fit_lm}
\title{internal helper function to setup a linear model for a given l1, l2, or l3 model}
\usage{
mobj_fit_lm(mobj = NULL, model_formula = NULL, data, id_cols = NULL)
}
\arguments{
\item{mobj}{a model object to be populated or modified}

\item{model_formula}{a formula of the model to be fit}

\item{data}{a data.frame containing all columns used in model fitting}

\item{id_cols}{a character vector of column names in \code{data} that identify the observations
and can be used for merging the model against related datasets}
}
\value{
a model object containing the fitted model
}
\description{
internal helper function to setup a linear model for a given l1, l2, or l3 model
}
\keyword{internal}
