% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glm_helper_functions.R
\name{mobj_refit_lm}
\alias{mobj_refit_lm}
\title{Helper function to re-calculate the l3 design matrix for a model based on available data}
\usage{
mobj_refit_lm(mobj, new_data)
}
\arguments{
\item{mobj}{a \code{hi_model_spec} object containing the L3 GLM model to run}

\item{data}{The run-level data frame containing data for all ids and sessions. This will be split into individual chunks}
}
\value{
a modified copy of \code{mobj} where the $by_subject field has been added
}
\description{
Helper function to re-calculate the l3 design matrix for a model based on available data
}
\details{
The function adds the $by_subject field, which contains the design matrices and contrasts
  for each subject and session in \code{data} based on the available data for that session. For example, if
  a subject is missing a few runs (or these are dropped from analysis), then some contrasts may change or drop out
  of the model.

The $by_subject field is a keyed data.table object containing list elements for the cope_list (mapping cope numbers
  to contrast names), the contrasts, and the design matrix for each session.
}
\keyword{internal}
