% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui_helpers.R
\name{ol_mean_min_max}
\alias{ol_mean_min_max}
\title{Round and format the mean, min, and max as a single character string}
\usage{
ol_mean_min_max(x, digits = 2, prespace = 0, showNA = "ifany", newline = TRUE)
}
\arguments{
\item{x}{the numeric variable to summarize}

\item{digits}{the number of digits to use in rounding}

\item{prespace}{An integer indicating the number of spaces to include at the beginning of the string. Default: 0}

\item{showNA}{whether to print information about NAs. As with \code{table}, the
options are "no", "ifany", and "always".}

\item{newline}{if TRUE, include a newline character \code{\n} at the end of the string}
}
\value{
a formatted character string with mean, min, and max
}
\description{
Round and format the mean, min, and max as a single character string
}
\keyword{internal}
