% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui_helpers.R
\name{ol_unique}
\alias{ol_unique}
\title{Print the top n unique values and frequencies of an atomic variable}
\usage{
ol_unique(
  x,
  max_show = 5,
  max_allowed = 25,
  prespace = 0,
  showNA = "ifany",
  newline = TRUE
)
}
\arguments{
\item{x}{the variable to tabulate}

\item{max_show}{the maximum number of unique values to show}

\item{max_allowed}{the maximum number of allowed unique values before the function summarizes as mean, min, max}

\item{prespace}{An integer indicating the number of spaces to include at the beginning of the string. Default: 0}

\item{showNA}{whether to print information about NAs. As with \code{table}, the
options are "no", "ifany", and "always".}

\item{newline}{if TRUE, include a newline character \code{\n} at the end of the string}
}
\value{
a formatted character string with frequencies for the top \code{max_show} categories
}
\description{
Print the top n unique values and frequencies of an atomic variable
}
\keyword{internal}
