% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glm_helper_functions.R
\name{pca_motion}
\alias{pca_motion}
\title{small helper function for compressing motion parameters using PCA}
\usage{
pca_motion(motion_df, num_pcs = 3L, zscore = TRUE, verbose = FALSE)
}
\arguments{
\item{motion_df}{volumes x motion parameters data frame for PCA compression}

\item{num_pcs}{number of principal components to extract}

\item{zscore}{whether to standardize motion parameters prior to PCA (this is a good idea)}

\item{verbose}{whether to print the variance explained by PCs and the multiple correlations of
among motion parameters}
}
\value{
A matrix of PCA-compressed motion regressors
}
\description{
small helper function for compressing motion parameters using PCA
}
\keyword{internal}
