% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmri_utility_fx.R
\name{place_dmat_on_time_grid}
\alias{place_dmat_on_time_grid}
\title{Function to convert dmat (runs x regressor list) to a time-oriented representation.
This yields a list of runs where each element is data.frame of volumes x regressors}
\usage{
place_dmat_on_time_grid(dmat, convolve = TRUE, run_timing = NULL, bdm_args)
}
\arguments{
\item{dmat}{A runs x regressors 2-d list where each element is a matrix containing
onset, duration, and value for a signal}

\item{convolve}{If \code{TRUE} (default), convolve the time-oriented signals with an HRF}

\item{run_timing}{A vector of cumulative start times for each run in a multi-run dataset}

\item{bdm_args}{A list of arguments passed to build_design_matrix, as well as a few fields added
during the initial argument parsing. See build_design_matrix for details. Should contain:
\itemize{
  \item convolve_wi_run TRUE/FALSE
  \item run_volumes Numeric vector of run length
  \item normalizations Character vector of HRF normalizations for each regressor.
    Options are "none", "durmax_1", or "evtmax_1".
  \item add_derivs A logical vector (\code{TRUE/FALSE}) of regressors whose temporal derivatives should
    be included. Temporal derivatives are only applied if \code{convolve} is \code{TRUE}.
  \item convmax_1 A logical vector (\code{TRUE/FALSE}) denoting whether to rescale max height to 1 after convolution
  \item high_pass The cutoff frequency (in Hz) used for high-pass filtering. If \code{NULL}, no filtering is applied.
  \item tr The repetition time (sometimes called TR) in seconds
  \item hrf_parameters The parameters for the double-gamma HRF
}}
}
\description{
Function to convert dmat (runs x regressor list) to a time-oriented representation.
This yields a list of runs where each element is data.frame of volumes x regressors
}
\author{
Michael Hallquist
}
\keyword{internal}
