% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glm_helper_functions.R
\name{populate_defaults}
\alias{populate_defaults}
\title{helper function to copy any missing fields in target}
\usage{
populate_defaults(target, defaults)
}
\arguments{
\item{target}{a named list to be populated by defaults if fields are missing}

\item{defaults}{a named list containing default values}
}
\description{
helper function to copy any missing fields in target
}
