% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glm_helper_functions.R
\name{populate_last_events}
\alias{populate_last_events}
\title{Internal function to add last_onset and last_offset columns to gpa$run_data based on events in l1_models}
\usage{
populate_last_events(gpa)
}
\arguments{
\item{gpa}{a \code{glm_pipeline_arguments} object containing valid $run_data and $l1_models objects}
}
\value{
a modified copy of gpa with $run_data populated with last_offset and last_onset columns (times in seconds)
}
\description{
Internal function to add last_onset and last_offset columns to gpa$run_data based on events in l1_models
}
\details{
The last_onset and last_offset columns are calculated for each run based on the timing of all
  events in the $l1_models$events list. These are then used to facilitate run truncation if the user
  requests truncation after a final onset or offset.
}
\keyword{internal}
