% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/job_tracking_functions.R
\name{populate_list_arg}
\alias{populate_list_arg}
\title{Internal helper function to update tracker_args object}
\usage{
populate_list_arg(list_to_populate, arg_name, new_value = NULL, append = FALSE)
}
\arguments{
\item{list_to_populate}{The list whose argument will be populated/updated}

\item{arg_name}{The named list element to update}

\item{new_value}{The new value to update the element with}

\item{append}{If TRUE, appends the new value to the current value using the paste function. Default: FALSE}
}
\description{
Internal helper function to update tracker_args object
}
\keyword{internal}
