% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_design_matrix_helpers.R
\name{prepare_signals_for_expansion}
\alias{prepare_signals_for_expansion}
\title{Prepare signals for expansion}
\usage{
prepare_signals_for_expansion(signals)
}
\arguments{
\item{signals}{list of signal specifications}
}
\value{
signals list with beta_series explicitly set
}
\description{
Ensures that beta_series is populated with TRUE or FALSE (default FALSE)
for each signal.
}
\keyword{internal}
