% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collinearity_diagnostics.R
\name{print.l1_collinearity_summary}
\alias{print.l1_collinearity_summary}
\title{Print method for l1_collinearity_summary objects}
\usage{
\method{print}{l1_collinearity_summary}(x, ...)
}
\arguments{
\item{x}{An l1_collinearity_summary object}

\item{...}{Additional arguments (ignored)}
}
\description{
Print method for l1_collinearity_summary objects
}
