% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fsl_helper_functions.R
\name{read_feat_dir}
\alias{read_feat_dir}
\title{helper function to look up core stats outputs from a .feat folder}
\usage{
read_feat_dir(
  feat_dir,
  what = "all",
  statistics = c("cope", "varcope", "zstat", "tstat"),
  include_missing = FALSE
)
}
\arguments{
\item{feat_dir}{a .feat folder containing the outputs of an FSL analysis}

\item{what}{which parts of the .feat folder should be parsed. If "all", extract everything.}

\item{statistics}{statistic image families to discover. The returned object
retains all existing statistic fields; unrequested fields contain aligned
\code{NA} values.}

\item{include_missing}{if \code{TRUE}, include expected statistic paths when
the corresponding image does not yet exist.}
}
\value{
a list containing sorted vectors of each stat output
}
\description{
helper function to look up core stats outputs from a .feat folder
}
