% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fsl_helper_functions.R
\name{read_feat_inputs}
\alias{read_feat_inputs}
\title{This is a small helper function that returns the inputs provided in the feat_files field for
a set of .fsf files.}
\usage{
read_feat_inputs(feat_list, recursive = FALSE)
}
\arguments{
\item{feat_list}{a vector of design.fsf filenames, and/or .feat/.gfeat directory names}

\item{recursive}{a boolean indicating whether to drill down and find lower-level inputs for .gfeat/.feat inputs}
}
\description{
This is a small helper function that returns the inputs provided in the feat_files field for
a set of .fsf files.
}
\details{
One can also pass in .feat or .gfeat directories and the function will
use the design.fsf files within each of these to find the inputs
}
\author{
Michael Hallquist
}
