% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fsl_helper_functions.R
\name{read_gfeat_dir}
\alias{read_gfeat_dir}
\title{helper function to look up core stats outputs from a .gfeat folder}
\usage{
read_gfeat_dir(
  gfeat_dir,
  what = "all",
  statistics = c("cope", "varcope", "zstat", "tstat"),
  include_missing = FALSE
)
}
\arguments{
\item{gfeat_dir}{a .gfeat folder containing the outputs of an FSL analysis}

\item{what}{What to parse in each folder. Currently just passed through to read_feat_dir}

\item{statistics}{statistic image families to discover in each cope directory.}

\item{include_missing}{if \code{TRUE}, include expected statistic paths when
the corresponding image does not yet exist.}
}
\value{
a list containing sorted vectors of each stat output
}
\description{
helper function to look up core stats outputs from a .gfeat folder
}
