% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/job_tracking_functions.R
\name{reset_tracking_sqlite_db}
\alias{reset_tracking_sqlite_db}
\title{Internal helper function to reset tracking SQLite database}
\usage{
reset_tracking_sqlite_db(sqlite_db)
}
\arguments{
\item{sqlite_db}{Path to SQLite database used for tracking}
}
\description{
Internal helper function to reset tracking SQLite database
}
\keyword{internal}
