% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rle_dt.R
\name{rle_dt}
\alias{rle_dt}
\title{R6 class for a keyed data.table object that uses run-length encoding to reduce storage size}
\description{
R6 class for a keyed data.table object that uses run-length encoding to reduce storage size
}
\section{Methods}{
\subsection{Public methods}{
  \itemize{
    \item \href{#method-rle_dt-initialize}{\code{rle_dt$new()}}
    \item \href{#method-rle_dt-get}{\code{rle_dt$get()}}
    \item \href{#method-rle_dt-clone}{\code{rle_dt$clone()}}
  }
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-rle_dt-initialize"></a>}}
\if{latex}{\out{\hypertarget{method-rle_dt-initialize}{}}}
\subsection{\code{rle_dt$new()}}{
  Create an RLE-encoded copy of the data
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{rle_dt$new(data, keys = NULL, optimize_order = (length(keys) <= 4))}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{data}}{A data.frame or data.table object containing original data}
      \item{\code{keys}}{A character vector of keys within \code{data} that should be RLE-encoded}
      \item{\code{optimize_order}}{A boolean indicating whether to search for the smallest encoding scheme.
The default it to search in data with 4 or fewer keys, but not to search for 5+.
Optionally, if a positive integer, randomly test ordering over this number of permutations.}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-rle_dt-get"></a>}}
\if{latex}{\out{\hypertarget{method-rle_dt-get}{}}}
\subsection{\code{rle_dt$get()}}{
  Simple method to return the data.table with all columns in their original form.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{rle_dt$get()}
    \if{html}{\out{</div>}}
  }
  \subsection{Details}{
    Note that the data are modified slightly in that the keys columns are placed first,
  and the data are ordered in the order of the keys (as originally provided, left-to-right)
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-rle_dt-clone"></a>}}
\if{latex}{\out{\hypertarget{method-rle_dt-clone}{}}}
\subsection{\code{rle_dt$clone()}}{
  The objects of this class are cloneable with this method.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{rle_dt$clone(deep = FALSE)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{deep}}{Whether to make a deep clone.}
    }
    \if{html}{\out{</div>}}
  }
}

}
