% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_3dlmer_sepjobs.R
\name{run_3dlmer_sepjobs}
\alias{run_3dlmer_sepjobs}
\title{function to submit 3dLMEr jobs on a cluster}
\usage{
run_3dlmer_sepjobs(
  gpa,
  level = 3L,
  model_names = NULL,
  rerun = FALSE,
  wait_for = ""
)
}
\arguments{
\item{gpa}{a \code{glm_pipeline_arguments} object}

\item{level}{GLM level to run; currently only level 3 is supported.}

\item{model_names}{optional subset of L3 model names to run}

\item{rerun}{logical, whether to rerun existing models}

\item{wait_for}{job ID(s) to wait for}
}
\value{
vector of submitted job IDs
}
\description{
function to submit 3dLMEr jobs on a cluster
}
