% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glm_helper_functions.R
\name{run_afni_command}
\alias{run_afni_command}
\title{Wrapper for running an AFNI command safely within R}
\usage{
run_afni_command(
  args,
  afnidir = NULL,
  stdout = NULL,
  stderr = NULL,
  echo = TRUE,
  omp_num_threads = 1L,
  ...
)
}
\arguments{
\item{args}{AFNI command string to be run}

\item{afnidir}{Location of AFNI installation. If NULL, the function will search the environment for AFNIDIR or afni on PATH}

\item{stdout}{File target for redirecting stdout. If NULL, stdout will not be written to file.}

\item{stderr}{File target for redirecting stderr. If NULL, stderr will not be written to file.}

\item{echo}{Whether to print AFNI command to the screen. Default: TRUE}

\item{omp_num_threads}{sets the number of OpenMP threads used for this AFNI command (if supported)}

\item{...}{Arguments passed through to the `system` command.}
}
\value{
The exit status of the executed AFNI command. 0 for success, non-zero for failure
}
\description{
Wrapper for running an AFNI command safely within R
}
\details{
This command ensures that AFNI commands are run in an environment that is setup correctly
}
\examples{

\dontrun{
run_afni_command("3dcopy test_data copy_data")
}
}
\author{
Michael Hallquist
}
