% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_feat_sepjobs.R
\name{run_feat_sepjobs}
\alias{run_feat_sepjobs}
\title{function to submit a set of jobs on a cluster to estimate many Feat level 1 models}
\usage{
run_feat_sepjobs(
  gpa,
  level = 1L,
  model_names = NULL,
  rerun = FALSE,
  wait_for = ""
)
}
\arguments{
\item{gpa}{a \code{glm_pipeline_arguments} object containing model specification}

\item{model_names}{an optional level 1 model name used to subset the runs to submit to feat.
If not provided, all level 1 models in \code{gpa} will be submitted for feat estimation.}

\item{rerun}{a logical indicating whether to re-run an existing directory. Default: FALSE}

\item{wait_for}{a parent job that should complete before these jobs commence}
}
\value{
a vector of job ids for all files that were submitted to the cluster
}
\description{
function to submit a set of jobs on a cluster to estimate many Feat level 1 models
}
