% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glm_helper_functions.R
\name{run_fsl_command}
\alias{run_fsl_command}
\title{Wrapper for running an FSL command safely within R}
\usage{
run_fsl_command(
  args,
  fsldir = NULL,
  stdout = NULL,
  stderr = NULL,
  echo = TRUE,
  ...
)
}
\arguments{
\item{args}{FSL command string to be run}

\item{fsldir}{Location of FSL installation. If NULL, the function will search the environment for FSLDIR or FSL commands in the PATH}

\item{stdout}{File target for redirecting stdout. If NULL, stdout will not be captured}

\item{stderr}{File target for redirecting stderr. If NULL, stderr will not be captured}

\item{echo}{Whether to print FSL command to the screen. Default: TRUE}

\item{...}{Arguments passed through to the `system` command.}
}
\value{
The exit status of the executed FSL command. 0 for success, non-zero for failure
}
\description{
Wrapper for running an FSL command safely within R
}
\details{
This command ensures that FSL command are run in an environment with FSL setup correctly.
}
\examples{
\dontrun{
run_fsl_command("fslmaths test_data copy_data")
}
}
\author{
Michael Hallquist
}
