% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_glm_pipeline.R
\name{run_glm_pipeline}
\alias{run_glm_pipeline}
\title{primary function for running a GLM analysis pipeline}
\usage{
run_glm_pipeline(
  gpa,
  l1_model_names = "prompt",
  l2_model_names = NULL,
  l3_model_names = NULL,
  glm_software = NULL
)
}
\arguments{
\item{gpa}{a glm_pipeline_arguments object containing a model specification (created by setup_glm_pipeline)}

\item{l1_model_names}{a character vector of level 1 model names (specified during build_l1_models) that should be executed}

\item{l2_model_names}{a character vector of level 2 model names (specified during build_l2_models) that should be executed}

\item{l3_model_names}{a character vector of level 3 model names (specified during build_l3_models) that should be executed}

\item{glm_software}{which glm software should be used for model estimation (not implemented yet)}
}
\description{
primary function for running a GLM analysis pipeline
}
