% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_spm_sepjobs.R
\name{run_spm_sepjobs}
\alias{run_spm_sepjobs}
\title{Submit SPM jobs to the cluster for unattended MATLAB/Octave execution}
\usage{
run_spm_sepjobs(
  gpa,
  level = 1L,
  model_names = NULL,
  rerun = FALSE,
  wait_for = ""
)
}
\arguments{
\item{gpa}{a \code{glm_pipeline_arguments} object containing model specification}

\item{level}{level of analysis (1 or 3)}

\item{model_names}{optional model names used to subset SPM runs}

\item{rerun}{logical indicating whether to rerun existing directories}

\item{wait_for}{optional parent job ids that should complete before these jobs commence}
}
\value{
a vector of job ids for all scripts that were submitted
}
\description{
Submit SPM jobs to the cluster for unattended MATLAB/Octave execution
}
