% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/afni_3dlmer.R
\name{sanitize_3dlmer_model_formula}
\alias{sanitize_3dlmer_model_formula}
\title{Normalize an AFNI 3dLMEr model formula}
\usage{
sanitize_3dlmer_model_formula(model_formula)
}
\arguments{
\item{model_formula}{fixed-effects formula string.}
}
\value{
A character string with whitespace removed.
}
\description{
Normalize an AFNI 3dLMEr model formula
}
\keyword{internal}
