% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glm_helper_functions.R
\name{sched_args_to_header}
\alias{sched_args_to_header}
\title{helper function to convert the $sched_args field to a vector
 of directives that can be included dynamically in the header of
 PBS or SBATCH scripts.}
\usage{
sched_args_to_header(gpa)
}
\arguments{
\item{gpa}{a \code{glm_pipeline_arguments} object containing the
$parallel$sched_args field}
}
\value{
a character vector where the $sched_args elements are converted to
  corresponding # SBATCH or # PBS directives for inclusion in scripts
}
\description{
helper function to convert the $sched_args field to a vector
 of directives that can be included dynamically in the header of
 PBS or SBATCH scripts.
}
\keyword{internal}
