% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glm_helper_functions.R
\name{setdiff_list_combn}
\alias{setdiff_list_combn}
\title{internal function to print bidirectional set differences in a list containing vectors to be compared}
\usage{
setdiff_list_combn(l)
}
\arguments{
\item{l}{a named list containing vectors to be compared}
}
\value{
NULL (invisibly)
}
\description{
internal function to print bidirectional set differences in a list containing vectors to be compared
}
\details{
All pairwise combinations of vectors in the list will be compared. The setdiff() operation is
  run twice for each pair, reflecting the two directions of comparison (s1 %notin% s2 and s2 %notin% s1).
}
\keyword{internal}
