% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_glm_pipeline.R
\name{setup_compute_environment}
\alias{setup_compute_environment}
\title{Helper function for setting up and modifying the compute environment for scripts generated by the pipeline}
\usage{
setup_compute_environment(gpa)
}
\arguments{
\item{gpa}{a \code{glm_pipeline_arguments} object}
}
\value{
a modified gpa object containing settings for the compute environment
}
\description{
Helper function for setting up and modifying the compute environment for scripts generated by the pipeline
}
