% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_l2_models.R
\name{setup_l2_models}
\alias{setup_l2_models}
\title{This function generates the inputs for an FSL level 2 analysis, where multiple runs for a subject are combined using
fixed effects estimation.}
\usage{
setup_l2_models(gpa, l2_model_names = NULL, l1_model_names = NULL)
}
\arguments{
\item{gpa}{a \code{glm_pipeline_arguments} object containing analysis speecification}

\item{l2_model_names}{a subset of L2 models to be setup by this function. If not specified,
all models in gpa$l2_models will be included}

\item{l1_model_names}{a subset of L1 models to be passed to L2 by this function. If not
specified, all models in gpa$l1_models will be included}
}
\description{
This function generates the inputs for an FSL level 2 analysis, where multiple runs for a subject are combined using
fixed effects estimation.
}
\details{
This function will setup FSL level 2 (subject) .fsf files for all combinations of
  \code{l2_model_names} and \code{l1_model_names}.
}
\author{
Michael Hallquist
}
