% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_l3_models.R
\name{setup_l3_models}
\alias{setup_l3_models}
\title{This function generates the inputs for level 3 analyses, where multi-subject data are analyzed
  in group analyses.}
\usage{
setup_l3_models(
  gpa,
  l3_model_names = NULL,
  l2_model_names = NULL,
  l1_model_names = NULL
)
}
\arguments{
\item{gpa}{a \code{glm_pipeline_arguments} object containing analysis speecification}

\item{l3_model_names}{a subset of L3 models to be setup by this function. If not specified,
all models in gpa$l2_models will be included}

\item{l2_model_names}{a subset of L2 models to be setup by this function. If not specified,
all models in gpa$l2_models will be included}

\item{l1_model_names}{a subset of L1 models to be passed to L2 by this function. If not
specified, all models in gpa$l1_models will be included}
}
\description{
This function generates the inputs for level 3 analyses, where multi-subject data are analyzed
  in group analyses.
}
\details{
This function will setup FSL level 2 (subject) .fsf files for all combinations of
  \code{l2_model_names} and \code{l1_model_names}.

FSL 2-level versus 3-level setup

2-level setup (one run per subject)
  - Pass L1 .feat folders as input to L3 .fsf setup
  - In this approach, the copes in the .fsf pertain to the L1 cope numbers
  - Requires one .fsf per L3 model

3-level setup (multiple runs per subject, combined at L2)
  - Pass individual cope*.feat folders within subject .gfeat folders
  - The folder cope numbers pertain to L1 copes
  - The cope*.nii.gz in the cope*.feat subfolders pertain to the L2 contrasts
  - Requires one .fsf per L1 cope x L3 model combination
  - Example: FSL_L2.gfeat/cope3.feat/stats/cope1.nii.gz
     ==> cope3 is the third contrast in the L1 feat model
     ==> cope1 is the first contrast in the L2 feat model
}
\author{
Michael Hallquist
}
