% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_job_status.R
\name{show_job_status}
\alias{show_job_status}
\title{Function which queries for the status of jobs associated with the given gpa object}
\usage{
show_job_status(
  gpa = NULL,
  batch_id = NULL,
  days_back = 1,
  desc = TRUE,
  latest_only = TRUE,
  update = TRUE
)
}
\arguments{
\item{gpa}{you glm_pipeline_arguments list}

\item{batch_id}{filter by batch_id from run_glm_pipeline}

\item{days_back}{how far back in time to search for jobs (in days)}

\item{desc}{whether to sort the status by descending order (newest first)}

\item{latest_only}{whether to only show the latest status of each job or all records}

\item{update}{whether or not to update the status of unresolved jobs}
}
\value{
a data.frame containing the status of jobs
}
\description{
Function which queries for the status of jobs associated with the given gpa object
}
