% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{sigmoid}
\alias{sigmoid}
\title{Sigmoid transform}
\arguments{
\item{x}{value to be transformed by sigmoid}

\item{beta}{beta slope (steepness) of sigmoid transform}
}
\description{
Sigmoid transform
}
\keyword{internal}
