% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specify_contrasts.R
\name{specify_contrasts}
\alias{specify_contrasts}
\title{Internal function to add, edit, rename, delete contrasts}
\usage{
specify_contrasts(mobj = NULL, signals = NULL, spec_list = NULL)
}
\arguments{
\item{mobj}{\code{l1_model_spec} or \code{hi_model_spec} object}

\item{signals}{a list of signals that provide information about regressor specification}

\item{spec_list}{a list of settings for this model based on an external YAML/JSON specification file}
}
\description{
Internal function to add, edit, rename, delete contrasts
}
\author{
Michael Hallquist
}
\keyword{internal}
