% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spm_extract_anatomical_rois.R
\name{spm_extract_anatomical_rois}
\alias{spm_extract_anatomical_rois}
\title{This is a wrapper around the spm_extract_anatomical_rois.m script in the inst directory}
\usage{
spm_extract_anatomical_rois(
  l1spmdirs,
  masks,
  threshold = 0.2,
  threshdesc = "none",
  session = 1,
  extent = 0,
  adjust_F_index = 1,
  contrast_index = NULL,
  ncores = 1,
  spm_path = "/gpfs/group/mnh5174/default/lab_resources/spm12",
  matlab_path = "/opt/aci/sw/matlab/R2017b/bin"
)
}
\arguments{
\item{l1spmdirs}{character vector of level 1 SPM directories containing SPM.mat files}

\item{masks}{character vector of NIfTI mask images for each anatomical ROI of interest}

\item{threshold}{p-value threshold applied to contrast within mask before extraction}

\item{threshdesc}{multiple comparisons correction on p-value. 'none' or 'FWE'}

\item{session}{which session (run) to use for extracting time series}

\item{extent}{exclude clusters having fewer than voxels than extent}

\item{adjust_F_index}{index of F-test in SPM.mat to adjust for all effects of interest}

\item{contrast_index}{index of t-test contrast in SPM.mat that is of interest}

\item{ncores}{number of cores to use in a parallel approach; function parallelizes over \code{l1spmdirs}}

\item{spm_path}{path to spm12 installation; added to MATLAB path at runtime}

\item{matlab_path}{location of MATLAB binary; used with matlabr for run_matlab_code()}
}
\description{
This is a wrapper around the spm_extract_anatomical_rois.m script in the inst directory
}
\author{
Michael Hallquist
}
