% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spm_l1_model.R
\name{spm_l1_model}
\alias{spm_l1_model}
\title{Setup a first-level SPM model using a build_design_matrix object}
\usage{
spm_l1_model(
  id = NULL,
  session = NULL,
  l1_confound_files = NULL,
  d_obj,
  gpa,
  model_name = NULL,
  run_nifti = NULL,
  execute_spm = FALSE
)
}
\arguments{
\item{l1_confound_files}{optional character vector of nuisance regressor files,
one per run, to include in SPM's multi_reg field}

\item{d_obj}{a single-subject design matrix object generated by build_design_matrix}

\item{gpa}{a gpa (glm_pipeline_arguments) object generated by setup_glm_pipeline}

\item{model_name}{a string indicating the model name within \code{gpa} to setup. If
you wish to setup multiple models, this is handled upstream in setup_l1_models.R}

\item{run_nifti}{optional character vector of NIfTI filenames used in l1 analysis.
Prefer to pass these via the build_design_matrix object in $run_nifti}

\item{execute_spm}{whether to execute SPM setup/estimation immediately. Default: FALSE}
}
\description{
Setup a first-level SPM model using a build_design_matrix object
}
\author{
Michael Hallquist
}
