% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spm_l3_model.R
\name{spm_l3_model}
\alias{spm_l3_model}
\title{Setup a third-level SPM model using lower-level con images}
\usage{
spm_l3_model(l3_df = NULL, gpa, execute_spm = FALSE, model_type = NULL)
}
\arguments{
\item{l3_df}{a data.frame containing one row per subject with fields including
id, session, l1_model, l3_model, l1_cope_name, and con_file}

\item{gpa}{a gpa (glm_pipeline_arguments) object containing model specification}

\item{execute_spm}{whether to execute SPM setup/estimation immediately. Default: FALSE}

\item{model_type}{optional override for SPM L3 model type. Options:
"flexible_factorial", "mreg", "one_sample"}
}
\description{
Setup a third-level SPM model using lower-level con images
}
\author{
Michael Hallquist
}
