% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sqlite_functions.R
\name{sqlite_table_exists}
\alias{sqlite_table_exists}
\title{helper function to check if a table exists in an SQLite database}
\usage{
sqlite_table_exists(sqlite_db = NULL, table = NULL, busy_timeout = NULL)
}
\arguments{
\item{sqlite_db}{Character path to SQLite database}

\item{table}{Character string name of table}
}
\description{
helper function to check if a table exists in an SQLite database
}
\keyword{internal}
