% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/job_tracking_functions.R
\name{submit_tracking_query}
\alias{submit_tracking_query}
\title{Internal helper function to submit a query to the tracking SQLite database}
\usage{
submit_tracking_query(str, sqlite_db, param = NULL)
}
\arguments{
\item{str}{Character string SQLite query}

\item{sqlite_db}{Path to SQLite database used for tracking}

\item{param}{List of parameters/arguments to be used in query}
}
\description{
Internal helper function to submit a query to the tracking SQLite database
}
\keyword{internal}
