% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glm_helper_functions.R
\name{summarize_contrasts}
\alias{summarize_contrasts}
\title{Internal helper function for printing a summary of a contrast matrix}
\usage{
summarize_contrasts(cmat)
}
\arguments{
\item{cmat}{a contrast matrix where row names are the contrast names and column
names are the column names of the corresponding design matrix.}
}
\description{
Internal helper function for printing a summary of a contrast matrix
}
\keyword{internal}
