% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_funcs.R
\name{summarize_pipeline}
\alias{summarize_pipeline}
\title{Helper function to better summarize GPA object}
\usage{
summarize_pipeline(gpa)
}
\arguments{
\item{gpa}{the \code{gpa} object}
}
\value{
a readable summary of the given gpa object
}
\description{
Helper function to better summarize GPA object
}
