% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_config.R
\name{summarize_project_config}
\alias{summarize_project_config}
\title{helper to summarize the contents of a project configuration file}
\usage{
summarize_project_config(config_file = NULL)
}
\arguments{
\item{config_file}{the path to the configuration file}
}
\description{
helper to summarize the contents of a project configuration file
}
\keyword{internal}
