% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_functions.R
\name{summary.glm_pipeline_arguments}
\alias{summary.glm_pipeline_arguments}
\title{Helper function to better summarize GPA object}
\usage{
\method{summary}{glm_pipeline_arguments}(object, ...)
}
\arguments{
\item{object}{the \code{glm_pipeline_arguments} object}

\item{...}{ignored.}
}
\value{
a readable summary of the given gpa object
}
\description{
Helper function to better summarize GPA object
}
