% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_functions.R
\name{summary.l1_model_set_models}
\alias{summary.l1_model_set_models}
\title{Helper function to better summarize GPA object's l1_models' models}
\usage{
\method{summary}{l1_model_set_models}(object, ...)
}
\arguments{
\item{object}{the \code{l1_model_set_models} object}

\item{...}{ignored.}
}
\value{
a readable summary of the given gpa object's l1_models' models
}
\description{
Helper function to better summarize GPA object's l1_models' models
}
