% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_functions.R
\name{summary.l1_model_set_signals}
\alias{summary.l1_model_set_signals}
\title{Helper function to better summarize GPA object's l1_models' signals}
\usage{
\method{summary}{l1_model_set_signals}(sl)
}
\arguments{
\item{gpa}{the \code{gpa} object}
}
\value{
a readable summary of the given gpa object's l1_models' signals
}
\description{
Helper function to better summarize GPA object's l1_models' signals
}
