% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_l1_confounds.R
\name{test_exclude_run}
\alias{test_exclude_run}
\title{calculate whether to retain or exclude this run}
\usage{
test_exclude_run(
  gpa,
  id,
  session,
  run_number,
  generate_run_exclusion,
  all_confounds,
  lg,
  confound_path = NA_character_,
  motion_path = NA_character_
)
}
\arguments{
\item{gpa}{A \code{glm_pipeline_arguments} object.}

\item{id}{subject id.}

\item{session}{session number.}

\item{run_number}{run number.}

\item{generate_run_exclusion}{whether exclusion should be evaluated.}

\item{all_confounds}{Combined confound data.}

\item{lg}{Logger object.}

\item{confound_path}{Path to the raw confound file.}

\item{motion_path}{Path to the raw motion file.}
}
\description{
calculate whether to retain or exclude this run
}
