% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_l1_confounds.R
\name{test_generate_l1_confounds}
\alias{test_generate_l1_confounds}
\title{helper function to generate confounds file}
\usage{
test_generate_l1_confounds(gpa, analysis_outdir, run_number, lg)
}
\description{
helper function to generate confounds file
}
