% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_l1_confounds.R
\name{test_generate_run_exclusion}
\alias{test_generate_run_exclusion}
\title{helper function to generate run exclusion}
\usage{
test_generate_run_exclusion(gpa, id, session, run_number, l1_cached_df, lg)
}
\description{
helper function to generate run exclusion
}
