% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_l1_confounds.R
\name{test_generate_run_truncation}
\alias{test_generate_run_truncation}
\title{helper function to generate run truncation}
\usage{
test_generate_run_truncation(gpa, id, session, run_number, lg)
}
\description{
helper function to generate run truncation
}
