% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/job_tracking_functions.R
\name{tracking_df_to_tree}
\alias{tracking_df_to_tree}
\title{Helper to convert a tracking dataframe into a data.tree object}
\usage{
tracking_df_to_tree(tracking_df)
}
\arguments{
\item{tracking_df}{a data.frame object extracted from the job tracking SQLite databse}
}
\description{
Helper to convert a tracking dataframe into a data.tree object
}
\keyword{internal}
