% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_batch_job.R
\name{update_job_status}
\alias{update_job_status}
\title{record job failure in SQLite database}
\usage{
update_job_status(sqlite_db = NULL, jobid = NULL, status)
}
\arguments{
\item{sqlite_db}{character string specifying the SQLite database used for job tracking}

\item{jobid}{character string specifying the job id to update as failed}

\item{status}{character string specifying the job status to set. Must be one of: 
"QUEUED", "STARTED", "FAILED", "COMPLETED"}
}
\description{
record job failure in SQLite database
}
