% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_config.R
\name{update_project_config}
\alias{update_project_config}
\title{helper function to create/update project configuration json file in gpa output directory}
\usage{
update_project_config(gpa, job_sequence, sequence_id, batch_directory = NULL)
}
\arguments{
\item{gpa}{a glm_pipeline_arguments object}

\item{job_sequence}{a list with named logicals indicating whether 
that part of the pipeline was submitted or not
(currently finalize, l1, l2, l3, cleanup)}

\item{sequence_id}{the unique identifer used for this batch sequence submission}

\item{batch_directory}{a path to the batch directory}
}
\description{
helper function to create/update project configuration json file in gpa output directory
}
\keyword{internal}
