% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/job_tracking_functions.R
\name{update_tracked_job_status}
\alias{update_tracked_job_status}
\title{Update job status in tracking SQLite database}
\usage{
update_tracked_job_status(
  sqlite_db = NULL,
  job_id = NULL,
  status,
  cascade = FALSE,
  exclude = NULL
)
}
\arguments{
\item{sqlite_db}{Character string specifying the SQLite database used for job tracking}

\item{job_id}{Character string specifying the job id to update as failed}

\item{status}{Character string specifying the job status to set. Must be one of: 
"QUEUED", "STARTED", "FAILED", "COMPLETED", "FAILED_BY_EXT"}

\item{exclude}{Any job ids to ignore when cascading a status}
}
\description{
Update job status in tracking SQLite database
}
