% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_glm_pipeline.R
\name{validate_block_data}
\alias{validate_block_data}
\title{internal function to validate the contents of a block data.frame}
\usage{
validate_block_data(df)
}
\description{
internal function to validate the contents of a block data.frame
}
\keyword{internal}
