% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glm_helper_functions.R
\name{validate_dhms}
\alias{validate_dhms}
\title{helper function to validate format of walltime inputs for HPC submission}
\usage{
validate_dhms(str)
}
\arguments{
\item{str}{string containing a duration that may include a days specification}
}
\description{
helper function to validate format of walltime inputs for HPC submission
}
\keyword{internal}
