% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnose_pipeline.R
\name{view_log}
\alias{view_log}
\title{Function for quick viewing of log files in the command line}
\usage{
view_log(input, from_end = TRUE, lines = 15L, level = NULL, id = NULL)
}
\arguments{
\item{input}{A character path to a project folder or a gpa object.}

\item{from_end}{Logical. If TRUE, view lines from end of file instead of beginning.}

\item{lines}{An integer. How many lines from beginnning/end to view.}

\item{level}{An integer equal to 1, 2 or 3 (model level).}

\item{id}{An integer or NULL. The ID number if wishing to view individual-level log.}
}
\value{
invisible NULL
}
\description{
Function for quick viewing of log files in the command line
}
\author{
Zach Vig
}
