% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmri_utility_fx.R
\name{visualize_design_matrix}
\alias{visualize_design_matrix}
\title{Visualize design matrix, including event onset times and run boundaries}
\usage{
visualize_design_matrix(
  d,
  outfile = NULL,
  run_boundaries = NULL,
  events = NULL,
  include_baseline = FALSE
)
}
\arguments{
\item{d}{a concatenated design matrix created by \code{build_design_matrix} and passed to \code{concat_design_runs}}

\item{outfile}{a filename used to export the design matrix using \code{ggsave}}

\item{run_boundaries}{a named vector of positions in the time series where a run boundary occurs (used for plotting)}

\item{events}{a named list of vectors containing the times of each event}

\item{include_baseline}{whether to display the baseline regressors in the design.}
}
\description{
Visualize design matrix, including event onset times and run boundaries
}
