% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/afni_3dlmer.R
\name{write_3dlmer_files}
\alias{write_3dlmer_files}
\title{Helper to write 3dLMEr script and data table}
\usage{
write_3dlmer_files(output_dir, dt, cmd, script_name = "run_3dlmer.sh")
}
\arguments{
\item{output_dir}{directory where files should be written}

\item{dt}{the data table data.frame}

\item{cmd}{the 3dLMEr command string}

\item{script_name}{name of the shell script to write}
}
\value{
list with paths to script and data table
}
\description{
Helper to write 3dLMEr script and data table
}
\keyword{internal}
