% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/afni_clustsim.R
\name{afni_3dclustsim}
\alias{afni_3dclustsim}
\title{R6 class for 3dClustSim automation}
\description{
R6 class for 3dClustSim automation

R6 class for 3dClustSim automation
}
\keyword{internal}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{fwhmx_set}}{A afni_3dfwhmx_list object containing 3dFWHMx information for all fwhmx_input_files}

\item{\code{inset_files}}{A character vector of files to use directly as volumes to threshold and clusterize}

\item{\code{insdat_file}}{An sdat file containing permutations to be passed to 3dClustSim through -insdat}

\item{\code{insdat_file}}{A mask file corresponding to the insdat_file data that indicates where each value is in space}

\item{\code{only}}{used if residuals_file is passed in, this contains the object for running the permutations
return TRUE/FALSE for whether this clustsim relies on the ACF estimates from 3dFWHMx
Simple method to refresh the clustsim_df, fwhmx files, and 3dttest++ permutation files
This is useful if the object needs to be updated just in time to determine whether permutations or ACF params
are available}
}
\if{html}{\out{</div>}}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{only}}{used if residuals_file is passed in, this contains the object for running the permutations
return TRUE/FALSE for whether this clustsim relies on the ACF estimates from 3dFWHMx
Simple method to refresh the clustsim_df, fwhmx files, and 3dttest++ permutation files
This is useful if the object needs to be updated just in time to determine whether permutations or ACF params
are available}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{afni_3dclustsim$new()}}
\item \href{#method-submit}{\code{afni_3dclustsim$submit()}}
\item \href{#method-get_clustsim_df}{\code{afni_3dclustsim$get_clustsim_df()}}
\item \href{#method-get_clustsim_output_files}{\code{afni_3dclustsim$get_clustsim_output_files()}}
\item \href{#method-get_call}{\code{afni_3dclustsim$get_call()}}
\item \href{#method-get_out_dir}{\code{afni_3dclustsim$get_out_dir()}}
\item \href{#method-get_ncpus}{\code{afni_3dclustsim$get_ncpus()}}
\item \href{#method-use_fwhmx_acf}{\code{afni_3dclustsim$use_fwhmx_acf()}}
\item \href{#method-is_complete}{\code{afni_3dclustsim$is_complete()}}
\item \href{#method-refresh}{\code{afni_3dclustsim$refresh()}}
\item \href{#method-clone}{\code{afni_3dclustsim$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{afni_3dclustsim$new(
  out_dir = NULL,
  prefix = NULL,
  fwhmx_input_files = NULL,
  fwhmx_mask_files = NULL,
  residuals_file = NULL,
  residuals_mask_file = NULL,
  residuals_njobs = NULL,
  inset_files = NULL,
  insdat_file = NULL,
  insdat_mask_file = NULL,
  dxyz = NULL,
  nxyz = NULL,
  clustsim_mask = NULL,
  acf_params = NULL,
  nopad = NULL,
  pthr = NULL,
  athr = NULL,
  iter = NULL,
  nodec = NULL,
  seed = NULL,
  scheduler = NULL,
  ncpus = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-submit"></a>}}
\if{latex}{\out{\hypertarget{method-submit}{}}}
\subsection{Method \code{submit()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{afni_3dclustsim$submit(force = FALSE)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_clustsim_df"></a>}}
\if{latex}{\out{\hypertarget{method-get_clustsim_df}{}}}
\subsection{Method \code{get_clustsim_df()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{afni_3dclustsim$get_clustsim_df()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_clustsim_output_files"></a>}}
\if{latex}{\out{\hypertarget{method-get_clustsim_output_files}{}}}
\subsection{Method \code{get_clustsim_output_files()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{afni_3dclustsim$get_clustsim_output_files()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_call"></a>}}
\if{latex}{\out{\hypertarget{method-get_call}{}}}
\subsection{Method \code{get_call()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{afni_3dclustsim$get_call()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_out_dir"></a>}}
\if{latex}{\out{\hypertarget{method-get_out_dir}{}}}
\subsection{Method \code{get_out_dir()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{afni_3dclustsim$get_out_dir()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_ncpus"></a>}}
\if{latex}{\out{\hypertarget{method-get_ncpus}{}}}
\subsection{Method \code{get_ncpus()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{afni_3dclustsim$get_ncpus()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-use_fwhmx_acf"></a>}}
\if{latex}{\out{\hypertarget{method-use_fwhmx_acf}{}}}
\subsection{Method \code{use_fwhmx_acf()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{afni_3dclustsim$use_fwhmx_acf()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-is_complete"></a>}}
\if{latex}{\out{\hypertarget{method-is_complete}{}}}
\subsection{Method \code{is_complete()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{afni_3dclustsim$is_complete()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-refresh"></a>}}
\if{latex}{\out{\hypertarget{method-refresh}{}}}
\subsection{Method \code{refresh()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{afni_3dclustsim$refresh()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{afni_3dclustsim$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
