% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/afni_3dclustsim.R
\name{afni_3dclustsim_list}
\alias{afni_3dclustsim_list}
\title{R6 class for a list of 3dClustSim runs}
\description{
R6 class for a list of 3dClustSim runs

R6 class for a list of 3dClustSim runs
}
\keyword{internal}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-afni_3dclustsim_list-new}{\code{afni_3dclustsim_list$new()}}
\item \href{#method-afni_3dclustsim_list-submit}{\code{afni_3dclustsim_list$submit()}}
\item \href{#method-afni_3dclustsim_list-get_objs}{\code{afni_3dclustsim_list$get_objs()}}
\item \href{#method-afni_3dclustsim_list-clone}{\code{afni_3dclustsim_list$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-afni_3dclustsim_list-new"></a>}}
\if{latex}{\out{\hypertarget{method-afni_3dclustsim_list-new}{}}}
\subsection{Method \code{new()}}{
create a new afni_3dclustsim_list object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{afni_3dclustsim_list$new(obj_list = NULL, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-afni_3dclustsim_list-submit"></a>}}
\if{latex}{\out{\hypertarget{method-afni_3dclustsim_list-submit}{}}}
\subsection{Method \code{submit()}}{
submit all jobs in this list
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{afni_3dclustsim_list$submit(force = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{force}}{If TRUE, pass force = TRUE to all the submit method of all subsidiary 3dClustSim objects}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-afni_3dclustsim_list-get_objs"></a>}}
\if{latex}{\out{\hypertarget{method-afni_3dclustsim_list-get_objs}{}}}
\subsection{Method \code{get_objs()}}{
return the clustsim objects
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{afni_3dclustsim_list$get_objs()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-afni_3dclustsim_list-clone"></a>}}
\if{latex}{\out{\hypertarget{method-afni_3dclustsim_list-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{afni_3dclustsim_list$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
