% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/afni_3dclustsim.R
\name{afni_3dclustsim_list}
\alias{afni_3dclustsim_list}
\title{R6 class for a list of 3dClustSim runs}
\description{
R6 class for a list of 3dClustSim runs

R6 class for a list of 3dClustSim runs
}
\keyword{all}
\keyword{in}
\keyword{internal}
\keyword{jobs}
\keyword{list}
\keyword{submit}
\keyword{this}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{afni_3dclustsim_list$new()}}
\item \href{#method-submit}{\code{afni_3dclustsim_list$submit()}}
\item \href{#method-get_objs}{\code{afni_3dclustsim_list$get_objs()}}
\item \href{#method-clone}{\code{afni_3dclustsim_list$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{afni_3dclustsim_list$new(obj_list = NULL, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-submit"></a>}}
\if{latex}{\out{\hypertarget{method-submit}{}}}
\subsection{Method \code{submit()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{afni_3dclustsim_list$submit(force = FALSE)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_objs"></a>}}
\if{latex}{\out{\hypertarget{method-get_objs}{}}}
\subsection{Method \code{get_objs()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{afni_3dclustsim_list$get_objs()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{afni_3dclustsim_list$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
