% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/afni_3dfwhmx_list.R
\name{afni_3dfwhmx_list}
\alias{afni_3dfwhmx_list}
\title{R6 class for running 3dFWHMx on a group of input files using a scheduler/cluster}
\description{
R6 class for running 3dFWHMx on a group of input files using a scheduler/cluster

R6 class for running 3dFWHMx on a group of input files using a scheduler/cluster
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{afni_3dfwhmx_list$new()}}
\item \href{#method-refresh}{\code{afni_3dfwhmx_list$refresh()}}
\item \href{#method-submit}{\code{afni_3dfwhmx_list$submit()}}
\item \href{#method-get_batch}{\code{afni_3dfwhmx_list$get_batch()}}
\item \href{#method-get_acf_average}{\code{afni_3dfwhmx_list$get_acf_average()}}
\item \href{#method-get_acf_df}{\code{afni_3dfwhmx_list$get_acf_df()}}
\item \href{#method-get_effective_fwhm}{\code{afni_3dfwhmx_list$get_effective_fwhm()}}
\item \href{#method-is_complete}{\code{afni_3dfwhmx_list$is_complete()}}
\item \href{#method-clone}{\code{afni_3dfwhmx_list$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{afni_3dfwhmx_list$new(
  input_files = NULL,
  mask_files = NULL,
  scheduler = NULL,
  wall_time = "10:00:00",
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-refresh"></a>}}
\if{latex}{\out{\hypertarget{method-refresh}{}}}
\subsection{Method \code{refresh()}}{
recreate afni_3dfwhmx objects and completion status
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{afni_3dfwhmx_list$refresh(...)}\if{html}{\out{</div>}}
}

\subsection{Details}{
useful for updating job status and ACF params after a run completes
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-submit"></a>}}
\if{latex}{\out{\hypertarget{method-submit}{}}}
\subsection{Method \code{submit()}}{
Submit the 3dFWHMx batch for all inputs to the scheduler
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{afni_3dfwhmx_list$submit(force = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{force}}{If \code{TRUE}, completed 3dFWHMx runs will be included in the batch. Default: FALSE}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_batch"></a>}}
\if{latex}{\out{\hypertarget{method-get_batch}{}}}
\subsection{Method \code{get_batch()}}{
Return R_batch_job objects used to submit 3dFWHMx for all inputs
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{afni_3dfwhmx_list$get_batch(include_complete = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{include_complete}}{If TRUE, complete 3dFWHMx jobs will be included in the batch, leading these to be re-run}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_acf_average"></a>}}
\if{latex}{\out{\hypertarget{method-get_acf_average}{}}}
\subsection{Method \code{get_acf_average()}}{
method to calculate the overall ACF average across all input datasets
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{afni_3dfwhmx_list$get_acf_average(allow_incomplete_fwhmx = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{allow_incomplete_fwhmx}}{If \code{TRUE}, the average will be calculated even if some datasets
do not have 3dFWHMx parameter estimates (e.g., if they crashed or were never run). Default. \code{FALSE}.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
a three-element vector containing the ACF estimates averaged across all datasets
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_acf_df"></a>}}
\if{latex}{\out{\hypertarget{method-get_acf_df}{}}}
\subsection{Method \code{get_acf_df()}}{
return the 3dFWHMx results as a data.frame
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{afni_3dfwhmx_list$get_acf_df()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_effective_fwhm"></a>}}
\if{latex}{\out{\hypertarget{method-get_effective_fwhm}{}}}
\subsection{Method \code{get_effective_fwhm()}}{
return the effective FWHM estimated by 3dFWHMx -ACF
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{afni_3dfwhmx_list$get_effective_fwhm()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-is_complete"></a>}}
\if{latex}{\out{\hypertarget{method-is_complete}{}}}
\subsection{Method \code{is_complete()}}{
Simple method to return whether 3dFWHMx is complete for all input files
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{afni_3dfwhmx_list$is_complete()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
returns \code{TRUE} if 3dFWHMx has completed for all runs
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{afni_3dfwhmx_list$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
