% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/afni_whereami.R
\name{afni_whereami}
\alias{afni_whereami}
\title{wrapper class for AFNI whereami}
\description{
wrapper class for AFNI whereami

wrapper class for AFNI whereami
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{afni_whereami$new()}}
\item \href{#method-get_call}{\code{afni_whereami$get_call()}}
\item \href{#method-get_omask_call}{\code{afni_whereami$get_omask_call()}}
\item \href{#method-get_whereami_df}{\code{afni_whereami$get_whereami_df()}}
\item \href{#method-get_omask_df}{\code{afni_whereami$get_omask_df()}}
\item \href{#method-get_output_files}{\code{afni_whereami$get_output_files()}}
\item \href{#method-run}{\code{afni_whereami$run()}}
\item \href{#method-clone}{\code{afni_whereami$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{afni_whereami$new(
  afni_3dclusterize_obj = NULL,
  omask = NULL,
  coord_file = NULL,
  coord_file_columns = NULL,
  coord_vector = NULL,
  atlases = NULL,
  coord_orientation = NULL,
  coord_space = NULL,
  output_file = NULL,
  omask_output_file = NULL,
  afnidir = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_call"></a>}}
\if{latex}{\out{\hypertarget{method-get_call}{}}}
\subsection{Method \code{get_call()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{afni_whereami$get_call()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_omask_call"></a>}}
\if{latex}{\out{\hypertarget{method-get_omask_call}{}}}
\subsection{Method \code{get_omask_call()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{afni_whereami$get_omask_call()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_whereami_df"></a>}}
\if{latex}{\out{\hypertarget{method-get_whereami_df}{}}}
\subsection{Method \code{get_whereami_df()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{afni_whereami$get_whereami_df()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_omask_df"></a>}}
\if{latex}{\out{\hypertarget{method-get_omask_df}{}}}
\subsection{Method \code{get_omask_df()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{afni_whereami$get_omask_df()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_output_files"></a>}}
\if{latex}{\out{\hypertarget{method-get_output_files}{}}}
\subsection{Method \code{get_output_files()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{afni_whereami$get_output_files(exclude_missing = TRUE)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-run"></a>}}
\if{latex}{\out{\hypertarget{method-run}{}}}
\subsection{Method \code{run()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{afni_whereami$run(force = FALSE)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{afni_whereami$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
