% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_l1_models.R
\name{bl1_build_events}
\alias{bl1_build_events}
\title{helper function to build events consisting of onsets and durations}
\usage{
bl1_build_events(l1_model_set, trial_data)
}
\arguments{
\item{l1_model_set}{an l1_model_set object that may have extant events in it}

\item{trial_data}{the trial_data object from the \code{gpa} object}
}
\value{
a modified copy of l1_model_set with events added/updated
}
\description{
helper function to build events consisting of onsets and durations
}
\keyword{interanl}
