% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_l1_models.R
\name{bl1_build_signals}
\alias{bl1_build_signals}
\title{helper function to build level 1 signals}
\usage{
bl1_build_signals(
  l1_model_set,
  trial_data,
  block_data = NULL,
  subtrial_data = NULL,
  ppi_data = NULL,
  lg = NULL
)
}
\arguments{
\item{l1_model_set}{An \code{l1_model_set} object whose signals should be created or modified}

\item{trial_data}{A data.frame containing trial-level signal information}

\item{ppi_data}{An optional data.frame containing physiological signals for PPI analysis}
}
\value{
a modified version of \code{l1_model_set} with updated \code{$signals}
}
\description{
helper function to build level 1 signals
}
\keyword{internal}
