% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/afni_3dlmer.R
\name{build_3dlmer_command}
\alias{build_3dlmer_command}
\title{Internal function to construct the 3dLMEr command string}
\usage{
build_3dlmer_command(
  prefix,
  model_formula,
  qVars = NULL,
  glt_codes = NULL,
  data_table_file,
  mask = NULL,
  njobs = 1,
  ss_type = 3
)
}
\arguments{
\item{prefix}{output prefix for 3dLMEr}

\item{model_formula}{fixed effects formula string}

\item{qVars}{character vector of quantitative variables}

\item{glt_codes}{list of named GLT code strings}

\item{data_table_file}{path to the data table text file}

\item{mask}{path to the brain mask file}

\item{njobs}{number of parallel jobs for 3dLMEr}

\item{ss_type}{sum of squares type (default 3)}
}
\value{
a character string containing the 3dLMEr command
}
\description{
Internal function to construct the 3dLMEr command string
}
\keyword{internal}
