% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_l2_models.R
\name{build_l2_models}
\alias{build_l2_models}
\title{Interactive function to build an l1 model specification for setup_glm_pipeline}
\usage{
build_l2_models(gpa, regressor_cols = NULL)
}
\arguments{
\item{regressor_cols}{an optional character vector of columns in \code{data} that should be considered
as possible regressors}

\item{data}{a data.frame containing trial-level data for one or more subjects}

\item{model_set}{optional existing model_set to be modified}

\item{variable_mapping}{a vector of mappings between columns in \code{data} and internal constructs}
}
\value{
a \code{hi_model_set} object containing a list of higher-level regression models
}
\description{
Interactive function to build an l1 model specification for setup_glm_pipeline
}
\author{
Michael Hallquist
}
