% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glm_helper_functions.R, R/run_glm_pipeline.R
\name{choose_glm_models}
\alias{choose_glm_models}
\title{helper function to ask user to choose models at a given level for further processing}
\usage{
choose_glm_models(
  gpa,
  l1_model_names = NULL,
  l2_model_names = NULL,
  l3_model_names = NULL,
  lg = NULL
)

choose_glm_models(
  gpa,
  l1_model_names = NULL,
  l2_model_names = NULL,
  l3_model_names = NULL,
  lg = NULL
)
}
\arguments{
\item{gpa}{a \code{glm_pipeline_arguments} with models specified at a given level}

\item{l1_model_names}{a character vector of level 1 model names}

\item{l2_model_names}{a character vector of level 2 model names}

\item{l3_model_names}{a character vector of level 3 model names}

\item{model_names}{a user-specified string of models to process/use}

\item{level}{the level of GLM analysis to be specified (1, 2, or 3)}
}
\value{
a character vector of user-specified model names at this level

a named list containing all models that were selected along with additional information
  about whether to rerun existing models
}
\description{
helper function to ask user to choose models at a given level for further processing

helper function to guide user through process of choosing which models to run in GLM pipeline
}
\keyword{internal}
