% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fsl_helper_functions.R
\name{combine_feat_l3_to_afni}
\alias{combine_feat_l3_to_afni}
\title{Function to combine Feat L3 analyses into AFNI BRIK+HEAD files according to user-specified combinations}
\usage{
combine_feat_l3_to_afni(
  gpa,
  feat_l3_combined_filename = NULL,
  feat_l3_combined_briknames = NULL
)
}
\arguments{
\item{gpa}{a glm_pipeline_arguments object having a populated l3_model_setup field. The L3 analyses should also
be complete (i.e., after run_glm_pipeline).}

\item{feat_l3_combined_filename}{a glue expression for the path and filename prefix}

\item{feat_l3_combiend_briknames}{a glue expression for naming the subbriks in the AFNI output}
}
\description{
Function to combine Feat L3 analyses into AFNI BRIK+HEAD files according to user-specified combinations
}
