% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_l1_confounds.R
\name{confound_manipulations}
\alias{confound_manipulations}
\title{helper function to manipulate confounds}
\usage{
confound_manipulations(
  gpa,
  all_confounds,
  expected_l1_confound_file,
  run_df,
  demean,
  lg
)
}
\description{
helper function to manipulate confounds
}
