% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/event_lock_timeseries.R
\name{evt_align_decon_files}
\alias{evt_align_decon_files}
\title{Align a set of deconvolved time series files to an event of interest function.}
\usage{
evt_align_decon_files(
  d_files,
  trial_df,
  alignment = list(),
  aggregate_by = "atlas_value",
  tr = NULL,
  atlas_cuts = NULL,
  atlas_subset = NULL,
  out_file = NULL,
  ncpus = 8
)
}
\arguments{
\item{d_files}{A vector of deconvolved .csv.gz files created by \code{voxelwise_deconvolution}.}

\item{trial_df}{The trial-level data.frame containing id and run for each subject represented in \code{d_files}.}

\item{alignment}{A list containing alignment details passed to get_medusa_interpolated_ts}

\item{aggregate_by}{The column name in the individual deconvolved files used for averaging repeated units (e.g., voxels) into single
event-aligned time series. Most commonly, this is "atlas_value", which will lead to averaging of voxels within each parcel in the mask.}

\item{tr}{The repetition time of the sequence in seconds.}

\item{atlas_cuts}{For a continuous-valued atlas (e.g., containing a gradient of interest), a vector of cut points for binning values.
These cuts are applied to the \code{aggregate_by} column, commonly "atlas_value"}

\item{atlas_subset}{An optional numeric vector containing values of \code{aggregate_by}}

\item{mask}{A list of mask-related information, including ...}

\item{output_dir}{The output directory for the event-aligned csv file. If NULL, nothing it output}
}
\description{
Align a set of deconvolved time series files to an event of interest function.
}
\details{
This is intended to be used internally by \code{run_decon_alignment}, which accepts a set of mask/atlas files
  and alignments, then processes these in parallel.
}
