% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_glm_betas_atlas.R
\name{extract_glm_betas_atlas}
\alias{extract_glm_betas_atlas}
\title{Function to extract subject-level coefficients from an FSL group analysis}
\usage{
extract_glm_betas_atlas(
  gpa,
  mask_files,
  what = c("cope", "varcope", "zstat"),
  out_dir = getwd(),
  extract_l1 = "all",
  extract_l2 = "all",
  extract_l3 = "all",
  ncores = NULL,
  aggregate = TRUE,
  aggFUN = mean,
  return_data = TRUE,
  write_data = TRUE
)
}
\arguments{
\item{gpa}{a \code{glm_pipeline_arguments} object for which results of GLM analyses are already available}

\item{mask_files}{vector of filenames for NIfTI-based atlases. The function will loop over each and extract coefficients from
each unique value in each mask.}

\item{what}{which statistics to extract from each parcel. Default is cope (aka 'beta'), varcope, and zstat.}

\item{out_dir}{the directory to which statistics are written as .csv.gz files.}

\item{extract_l1}{a character vector of l1 models from which to extract run-level (level 1) statistics. If "none",
l1 statistic extraction will be skipped. If "all", then all l1 models will be extracted.}

\item{extract_l2}{a character vector of l2 models from which to extract subject-level (level 2) statistics. If "none",
l2 statistic extraction will be skipped. If "all", then all l2 models will be extracted (you will still be prompted
for which l1 models you want to extract from each l2 model).}

\item{extract_l3}{a character vector of l3 models from which to extract group-level (level 3) statistics. If "none",
l3 statistic extraction will be skipped. If "all", then all l3 models will be extracted (you will still be prompted
for which l1 and l2 models you want to extract from each l3 model).}

\item{aggregate}{whether to take the average (or other central tendency measure) of voxels within a given mask value. This
only pertains to integer-valued masks, not continuous ones.}

\item{aggFUN}{the function used to aggregate statistics for voxels within a given mask value. Default is mean.}

\item{return_data}{if TRUE, then extracted statistics will be returned as a list of data.frames with elements l1, l2, and l3.}

\item{write_data}{if TRUE, then extracted statistics will be written as .csv.gz files to \code{out_dir}.}
}
\description{
Function to extract subject-level coefficients from an FSL group analysis
}
