% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmri.pipeline-package.R
\docType{package}
\name{fmri.pipeline-package}
\alias{fmri.pipeline}
\alias{fmri.pipeline-package}
\title{fmri.pipeline: Analysis of fMRI Data on HPC clusters}
\description{
Setup of GLM and deconvolution analyses for large-scale analyses of fMRI data.
}
\author{
\strong{Maintainer}: Michael Hallquist \email{michael.hallquist@gmail.com}

}
\keyword{internal}
