\name{fmri.pipeline-package}
\alias{fmri.pipeline-package}
\alias{fmri.pipeline}
\docType{package}
\title{
\packageTitle{fmri.pipeline}
}
\description{
\packageDescription{fmri.pipeline}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{fmri.pipeline}
\packageIndices{fmri.pipeline}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{fmri.pipeline}

Maintainer: \packageMaintainer{fmri.pipeline}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
