% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmri_ts.R
\name{fmri_ts}
\alias{fmri_ts}
\title{R6 class representing a multivariate time series object for fMRI analysis}
\description{
R6 class representing a multivariate time series object for fMRI analysis

R6 class representing a multivariate time series object for fMRI analysis
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{ts_data}}{time x signals data.table}

\item{\code{ts_keys}}{RLE-encoded keying variables}

\item{\code{event_data}}{trial x event data, used for aligning time series with events}

\item{\code{vm}}{a list of variable mappings between internal constructs and input variable names}

\item{\code{tr}}{the repetition time (TR) of the fMRI sequence in seconds}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-fmri_ts-new}{\code{fmri_ts$new()}}
\item \href{#method-fmri_ts-get_ts}{\code{fmri_ts$get_ts()}}
\item \href{#method-fmri_ts-add_keys}{\code{fmri_ts$add_keys()}}
\item \href{#method-fmri_ts-replace_vm}{\code{fmri_ts$replace_vm()}}
\item \href{#method-fmri_ts-get_kvars}{\code{fmri_ts$get_kvars()}}
\item \href{#method-fmri_ts-get_vmvec}{\code{fmri_ts$get_vmvec()}}
\item \href{#method-fmri_ts-export}{\code{fmri_ts$export()}}
\item \href{#method-fmri_ts-clone}{\code{fmri_ts$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fmri_ts-new"></a>}}
\if{latex}{\out{\hypertarget{method-fmri_ts-new}{}}}
\subsection{Method \code{new()}}{
Create a new fmri_ts object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{fmri_ts$new(ts_data = NULL, event_data = NULL, vm = NULL, tr = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{ts_data}}{a data.frame or data.table containing time series}

\item{\code{event_data}}{a data.frame containing trial-level events that occurred in the time period represented by \code{ts_data}}

\item{\code{vm}}{a list of variable names used in \code{ts_data} and \code{event_data} that map onto internal constructs}

\item{\code{tr}}{the sampling rate (in seconds) of the fMRI data}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fmri_ts-get_ts"></a>}}
\if{latex}{\out{\hypertarget{method-fmri_ts-get_ts}{}}}
\subsection{Method \code{get_ts()}}{
method to get rehydrated time series object with key values
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{fmri_ts$get_ts(orig_names = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{orig_names}}{boolean indicating whether to return data.table with original naming scheme. Default: FALSE}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fmri_ts-add_keys"></a>}}
\if{latex}{\out{\hypertarget{method-fmri_ts-add_keys}{}}}
\subsection{Method \code{add_keys()}}{
method to add a variable in ts_data to the set of keying variables for further use
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{fmri_ts$add_keys(kv)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{kv}}{A vector of one or more variables to RLE-encode and add as keys the object}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fmri_ts-replace_vm"></a>}}
\if{latex}{\out{\hypertarget{method-fmri_ts-replace_vm}{}}}
\subsection{Method \code{replace_vm()}}{
method to replace one or more variable mappings in the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{fmri_ts$replace_vm(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{a set of arguments, each one of which replaces a field in
the variable mapping with a new specification}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fmri_ts-get_kvars"></a>}}
\if{latex}{\out{\hypertarget{method-fmri_ts-get_kvars}{}}}
\subsection{Method \code{get_kvars()}}{
return names of key variables
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{fmri_ts$get_kvars()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fmri_ts-get_vmvec"></a>}}
\if{latex}{\out{\hypertarget{method-fmri_ts-get_vmvec}{}}}
\subsection{Method \code{get_vmvec()}}{
return variable mapping information
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{fmri_ts$get_vmvec()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fmri_ts-export"></a>}}
\if{latex}{\out{\hypertarget{method-fmri_ts-export}{}}}
\subsection{Method \code{export()}}{
not currently used
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{fmri_ts$export(filename)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{filename}}{for writing out data}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fmri_ts-clone"></a>}}
\if{latex}{\out{\hypertarget{method-fmri_ts-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{fmri_ts$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
