% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fsl_l1_model.R
\name{fsl_l1_model}
\alias{fsl_l1_model}
\title{note: this is a small adapation from the original fslSCEPTICModel to avoid use of the clockfit objects and to move to the
simpler build_design_matrix approach and the use of the trial_statistics csv files from vba_fmri}
\usage{
fsl_l1_model(
  id = NULL,
  d_obj,
  gpa,
  model_name = NULL,
  mr_files = NULL,
  nvoxels = NULL,
  execute_feat = FALSE,
  lg = NULL
)
}
\arguments{
\item{d_obj}{a single-subject design matrix object generated by build_design_matrix}

\item{gpa}{a gpa (glm_pipeline_arguments) object generated by setup_glm_pipeline}

\item{model_name}{a string indicating the model name within \code{gpa} to setup. If
you wish to setup multiple models, this is handled upstream in setup_l1_models.R}

\item{mr_files}{an optional character vector of NIfTI filenames used in l1 analysis.
Prefer to pass these via the build_design_matrix object in $run_4d_files}
}
\description{
note: this is a small adapation from the original fslSCEPTICModel to avoid use of the clockfit objects and to move to the
simpler build_design_matrix approach and the use of the trial_statistics csv files from vba_fmri
}
\author{
Michael Hallquist
}
