% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fsl_l2_model.R
\name{fsl_l2_model}
\alias{fsl_l2_model}
\title{Estimate a level 2 (subject) model using FSL FEAT with fixed effects integration of runs}
\usage{
fsl_l2_model(l1_df = NULL, l2_model_name, gpa, execute_feat = FALSE)
}
\arguments{
\item{l1_df}{a data.frame containing all runs for a single subject and a single l1 model. This
data.frame defines the inputs for the L2 analysis (i.e., which runs to combine).}

\item{l2_model_name}{a model string in gpa$l2_models containing the L2 model to setup}

\item{gpa}{a \code{glm_pipeline_arguments} object containing model specification}

\item{execute_feat}{a logical indicating whether to run the L2 model after creating it}
}
\description{
Estimate a level 2 (subject) model using FSL FEAT with fixed effects integration of runs
}
\author{
Michael Hallquist
}
