% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_fwe_correction.R
\name{fwe_spec}
\alias{fwe_spec}
\title{R6 class for an FWE correction method that can apply to one or more models}
\description{
R6 class for an FWE correction method that can apply to one or more models

R6 class for an FWE correction method that can apply to one or more models
}
\keyword{internal}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-fwe_spec-new}{\code{fwe_spec$new()}}
\item \href{#method-fwe_spec-submit}{\code{fwe_spec$submit()}}
\item \href{#method-fwe_spec-clone}{\code{fwe_spec$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fwe_spec-new"></a>}}
\if{latex}{\out{\hypertarget{method-fwe_spec-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{fwe_spec$new(..., fwe_data = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fwe_spec-submit"></a>}}
\if{latex}{\out{\hypertarget{method-fwe_spec-submit}{}}}
\subsection{Method \code{submit()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{fwe_spec$submit()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fwe_spec-clone"></a>}}
\if{latex}{\out{\hypertarget{method-fwe_spec-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{fwe_spec$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
