% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glm_helper_functions.R
\name{generate_run_data_from_bids}
\alias{generate_run_data_from_bids}
\title{initial function to generate a run_data object from a BIDS-compliant folder}
\usage{
generate_run_data_from_bids(
  bids_dir,
  modality = "func",
  type = "task",
  task_name = "ridl",
  suffix = "_postprocessed"
)
}
\arguments{
\item{bids_dir}{a directory containing BIDS-compliant processed data for analysis}

\item{modality}{the subfolder within \code{bids_dir} that contains data of a certain modality.
Almost always 'func', which is the default.}

\item{type}{at present, always 'task' to denote this part of the BIDS filename... Not totally sure what else it could be}

\item{task_name}{the name of the task, which is appended with \code{type}}

\item{suffix}{an optional suffix in the expected filename (just before the file extension)}
}
\value{
a data.frame containing all run_nifti and confound_input_file results for subjects in the folder
}
\description{
initial function to generate a run_data object from a BIDS-compliant folder
}
\details{
The files should generally have a name like
  sub-220256_task-ridl3_space-MNI152NLin2009cAsym_desc-preproc_bold_postprocessed.nii.gz
  and be located in a folder like: /proj/mnhallqlab/proc_data/sub-220256/func/
  where 'func' is the \code{modality}, 'task' is the \code{type}, 'ridl' is the \code{task_name}, and
  '_postprocessed' is the \code{suffix}.
}
