% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glm_helper_functions.R
\name{get_confound_txt}
\alias{get_confound_txt}
\title{helper function to generate confounds txt file for inclusion as additional regressors}
\usage{
get_confound_txt(
  id = NULL,
  session = NULL,
  run_number = NULL,
  gpa,
  drop_volumes = 0L,
  last_volume = NULL,
  demean = TRUE
)
}
\arguments{
\item{id}{The subject id}

\item{session}{The session number}

\item{run_number}{The run number}

\item{gpa}{a \code{glm_pipeline_arguments} object containing pipeline specification}
}
\description{
helper function to generate confounds txt file for inclusion as additional regressors
}
