% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glm_helper_functions.R
\name{lookup_dim}
\alias{lookup_dim}
\title{Helper function to obtain the dimensions of a 4D file}
\usage{
lookup_dim(nifti)
}
\arguments{
\item{nifti}{a 4D nifti file}
}
\value{
a named vector of dimensions for the input (dim_x, dim_y, dim_z)
}
\description{
Helper function to obtain the dimensions of a 4D file
}
\details{
This function prefers to use fslval instead of an internal R library
  because both oro.nifti and RNifti are rather slow to obtain a single value from
  the NIfTI header
}
\keyword{internal}
