% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookup_nifti_inputs.R
\name{lookup_nifti_inputs}
\alias{lookup_nifti_inputs}
\title{helper function to locate nifti inputs for each run of data}
\usage{
lookup_nifti_inputs(gpa, add_run_volumes = TRUE, add_dim = TRUE)
}
\arguments{
\item{gpa}{a \code{glm_pipeline_arguments} object created by setup_glm_pipeline}

\item{add_run_volumes}{whether to lookup and add the number of volumes in each run_nifti in the run_volumes column}

\item{add_dim}{whether to lookup and add the dimensions of each run_nifti in the dim_x, dim_y, and dim_z columns.
This will also add an \code{nvoxels} column used in FEAT fsfs, which is x * y * z * t.}
}
\value{
a modified version of the \code{gpa} object that contains a $run_nifti field and relevant attribute columns (e.g., dim_x)
  the $run_data data.frame.
}
\description{
helper function to locate nifti inputs for each run of data
}
\author{
Michael Hallquist
}
\keyword{internal}
