% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixed_by.R
\name{mixed_by}
\alias{mixed_by}
\title{Mixed by runs a set of mixed-effects models for each combination of a set of factors. Its primary use is to
  run the same model on different splits of the data.}
\usage{
mixed_by(
  data,
  outcomes = NULL,
  rhs_model_formulae = NULL,
  split_on = NULL,
  external_df = NULL,
  external_merge_by = NULL,
  padjust_by = "term",
  padjust_method = "BY",
  outcome_transform = NULL,
  ncores = 1L,
  cl = NULL,
  refit_on_nonconvergence = 3,
  tidy_args = list(effects = "fixed", conf.int = TRUE),
  lmer_control = lmerControl(optimizer = "nloptwrap")
)
}
\arguments{
\item{data}{A data.frame or data.table object containing stacked data for each combination of the \code{split_on}
variables. The function will run separate mixed-effect models for each combination. Alternatively, a vector of filenames
can be passed, which will be read in sequentially and fit (.rds, .csv, .dat, and .txt supported at present).}

\item{outcomes}{A character vector of outcome variables to be analyzed}

\item{rhs_model_formulae}{A lme4-format formula specifying the exact model to be run for each data split.}

\item{split_on}{A character vector of columns in \code{data} used to split the analyses into separate models.}

\item{external_df}{An optional data.frame/data.table containing external data that should be joined with \code{data}
prior to model fitting. Useful if \code{data} contains external time series data and \code{external_df} is a
dataset of behavioral variables that do not vary by neural sensor/region. Optionally, this can be a single
filename to a .rds file if you want to pass in the filename, not the data itself.}

\item{external_merge_by}{A character vector specifying which columns of \code{external_df} and \code{data} should
be used for creating a combined dataset.}

\item{padjust_by}{A character vector or list consisting of one or more variables over which an adjusted p-value
should be calculated. This defaults to adjusting by each term (fixed effect) in the model, which will adjust for
all tests for a given term across variables in \code{split_on}.}

\item{padjust_method}{The adjustment method (see \code{?p.adjust}) for adjusting p-values. Multiple values
can be passed as a character vector, in which case multiple corrections will be added as distinct columns.}

\item{outcome_transform}{A vectorized function that will be applied to the outcome variable prior to running
the model. For example, \code{outcome_transform=function(x) { as.vector(scale(x)) } }.}

\item{ncores}{The number of compute cores to be used in the computation. Defaults to 1.}

\item{cl}{An optional external cl object (created by a variant of makeCluster) used for computation. Can
save the overhead of starting and stopping many workers in a loop context.}

\item{refit_on_nonconvergence}{The number of times a model should be refit if it does not converge. Final estimates
from one iteration are used as starting values for the next.}

\item{tidy_args}{A list of arguments passed to tidy.merMod for creating the coefficient data.frame. By default,
the function only returns the fixed effects and computes confidence intervals using the Wald method.}

\item{lmer_control}{An lmerControl object specifying any optimization settings to be passed to lmer()}
}
\value{
A data.table object containing all coefficients for each model estimated separately by \code{split_on}
}
\description{
Mixed by runs a set of mixed-effects models for each combination of a set of factors. Its primary use is to
  run the same model on different splits of the data.
}
