% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ptfce_spec.R
\name{ptfce_spec}
\alias{ptfce_spec}
\title{R6 class for pTFCE specification for one or more z-statistic images}
\description{
R6 class for pTFCE specification for one or more z-statistic images

R6 class for pTFCE specification for one or more z-statistic images
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{fwe_p}}{a vector of p-values used for familywise error (FWE) z-statistic threshold calculations in pTFCE.}

\item{\code{two_sided}}{It \code{TRUE}, run pTFCE on both tails of the statistic separately. If \code{FALSE},
only run pTFCE on the positve tail (z > 0).}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{ptfce_spec$new()}}
\item \href{#method-get_ptfce_calls}{\code{ptfce_spec$get_ptfce_calls()}}
\item \href{#method-get_expected_files}{\code{ptfce_spec$get_expected_files()}}
\item \href{#method-run}{\code{ptfce_spec$run()}}
\item \href{#method-submit}{\code{ptfce_spec$submit()}}
\item \href{#method-is_complete}{\code{ptfce_spec$is_complete()}}
\item \href{#method-get_clusters}{\code{ptfce_spec$get_clusters()}}
\item \href{#method-clone}{\code{ptfce_spec$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Create a new ptfce_spec object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ptfce_spec$new(
  gfeat_dir = NULL,
  zstat_numbers = NULL,
  fsl_smoothest_file = NULL,
  dof = NULL,
  residuals_file = NULL,
  z_files = NULL,
  mask_files = NULL,
  fwe_p = 0.05,
  two_sided = TRUE,
  write_thresh_imgs = TRUE,
  scheduler = NULL,
  time_per_zstat = NULL,
  memgb_per_command = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{gfeat_dir}}{One or more .gfeat folders containing a higher-level FSL analysis. These will be used for zstat images,
mask files, and fsl residual smoothness estimates.}

\item{\code{zstat_numbers}}{if a \code{gfeat_dir} is used, a vector of zstat numbers can also be provided to
subset the zstat images that are used in pTFCE correction. Ignored if gfeat_dir is not provided.}

\item{\code{fsl_smoothest_file}}{The smoothness file created by the smoothest command in FSL. Created by FEAT automatically
or can be run manually with smoothest on the res4d file. Used by ptfce to compute resels.}

\item{\code{dof}}{The degrees of freedom for the test of interest. Found in the dof file created by FSL and used by
ptfce for FWE correction.}

\item{\code{residuals_file}}{The residuals file from the group analysis used by ptfce for FWE correction. In FSL, this
is the res4d.nii.gz file created by FEAT.}

\item{\code{z_files}}{A vector of z-statistic filenames that should be corrected using pTFCE}

\item{\code{mask_files}}{A vector of mask filenames that correspond to \code{z_files}. If this is of length 1,
then the mask file will be recycled for all zstat images.}

\item{\code{fwe_p}}{A vector of p-values for which z-statistic thresholds will be calculated.}

\item{\code{two_sided}}{If TRUE, p-values for \code{fwe_p} are treated as two-tailed (i.e., p-values are divided by 2 in
the TFCE z-threshold calculation.)}

\item{\code{write_thresh_imgs}}{If TRUE, then pTFCE thresholds for each fwe_p will be applied to the TFCE image and saved
to the same folder as the z-statistic. These thresholded files let you look at the map at a given FWE threshold.}

\item{\code{scheduler}}{Which scheduler to use for submitting jobs. Options are 'local', 'slurm', and 'torque'.}

\item{\code{time_per_zstat}}{The amount of time to budget for each zstat to run through pTFCE in dd-hh:mm:ss format.
Default is 10:00 (10 minutes).}

\item{\code{memgb_per_command}}{How many GB of memory should be requested for each pTFCE command/run. If not provided, defaults to 8.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_ptfce_calls"></a>}}
\if{latex}{\out{\hypertarget{method-get_ptfce_calls}{}}}
\subsection{Method \code{get_ptfce_calls()}}{
method to return calls to external ptfce_zstat.R script for each zstat
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ptfce_spec$get_ptfce_calls(include_complete = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{include_complete}}{if TRUE, return calls for zstats that already appear to have
pTFCE-corrected images in place. Default: FALSE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_expected_files"></a>}}
\if{latex}{\out{\hypertarget{method-get_expected_files}{}}}
\subsection{Method \code{get_expected_files()}}{
return the vector of expected output files
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ptfce_spec$get_expected_files()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-run"></a>}}
\if{latex}{\out{\hypertarget{method-run}{}}}
\subsection{Method \code{run()}}{
Run pTFCE in this compute environment. This is not supported at present!
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ptfce_spec$run(force = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{force}}{if TRUE, re-run pTFCE on an existing output
method to submit all ptfce_zstat.R calls to a cluster based on the scheduler specified}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-submit"></a>}}
\if{latex}{\out{\hypertarget{method-submit}{}}}
\subsection{Method \code{submit()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ptfce_spec$submit(force = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{force}}{if TRUE, re-run pTFCE for zstat images that already appear to have pTFCE-corrected outputs in place}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-is_complete"></a>}}
\if{latex}{\out{\hypertarget{method-is_complete}{}}}
\subsection{Method \code{is_complete()}}{
returns \code{TRUE} if all expected pTFCE output files exist, \code{FALSE} if any output is missing
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ptfce_spec$is_complete()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_clusters"></a>}}
\if{latex}{\out{\hypertarget{method-get_clusters}{}}}
\subsection{Method \code{get_clusters()}}{
for each input file, obtain 3dClusterize objects that reflect the pTFCE-corrected clusters
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ptfce_spec$get_clusters(
  fwep = 0.05,
  clust_nvox = 10,
  NN = 1L,
  add_whereami = TRUE,
  whereami_atlases = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{fwep}}{the whole-brain familywise error rate to use (often .05)}

\item{\code{clust_nvox}}{The minimum number of voxels to allow in a given cluster}

\item{\code{NN}}{The cluster definition in AFNI terms. 1 = faces touch, 2 = edges touch, 3 = corners touch}

\item{\code{add_whereami}}{if TRUE, lookup labels for each cluster}

\item{\code{whereami_atlases}}{The atlases to request in the whereami lookup. If NULL, it uses the defaults}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
Note that even though pTFCE enhances clusters, you still see some very small clusters in some cases.
  Hence, there is no requirement to have clust_nvox > 1, but it may be a good idea for your sanity.
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ptfce_spec$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
