% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_fwe_correction.R
\name{ptfce_spec}
\alias{ptfce_spec}
\title{R6 class for pTFCE specification for one or more z-statistic images}
\description{
R6 class for pTFCE specification for one or more z-statistic images

R6 class for pTFCE specification for one or more z-statistic images
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{fwe_p}}{a vector of p-values used for familywise error (FWE) z-statistic threshold calculations in pTFCE.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{ptfce_spec$new()}}
\item \href{#method-get_ptfce_calls}{\code{ptfce_spec$get_ptfce_calls()}}
\item \href{#method-get_expected_files}{\code{ptfce_spec$get_expected_files()}}
\item \href{#method-run}{\code{ptfce_spec$run()}}
\item \href{#method-submit}{\code{ptfce_spec$submit()}}
\item \href{#method-all_expected_exist}{\code{ptfce_spec$all_expected_exist()}}
\item \href{#method-clone}{\code{ptfce_spec$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ptfce_spec$new(
  gfeat_dir = NULL,
  zstat_numbers = NULL,
  fsl_smoothest_file = NULL,
  dof = NULL,
  residuals_file = NULL,
  z_files = NULL,
  mask_files = NULL,
  fwe_p = 0.05,
  two_sided = TRUE,
  write_thresh_imgs = TRUE,
  scheduler = NULL,
  time_per_zstat = NULL,
  memgb_per_command = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{gfeat_dir}}{a .gfeat folder containing a higher-level FSL analysis. This will be used for zstat images
and mask files.}

\item{\code{zstat_numbers}}{if a \code{gfeat_dir} is used, a vector of zstat numbers can also be provided to
subset the zstat images that are used in pTFCE correction.}

\item{\code{z_files}}{A vector of z-statistic filenames that should be corrected using pTFCE}

\item{\code{mask_files}}{A vector of mask filenames that correspond to \code{z_files}. If this is of length 1,
then the mask file will be recycled for all zstat images.}

\item{\code{fwe_p}}{A vector of p-values for which z-statistic thresholds will be calculated.}

\item{\code{two_sided}}{If TRUE, p-values for \code{fwe_p} are treated as two-tailed (i.e., p-values are divided by 2 in
the TFCE z-threshold calculation.)}

\item{\code{write_thresh_imgs}}{If TRUE, then pTFCE thresholds for each fwe_p will be applied to the TFCE image and saved
to the same folder as the z-statistic. These thresholded files let you look at the map at a given FWE threshold.}

\item{\code{scheduler}}{Which scheduler to use for submitting jobs. Options are 'local', 'slurm', and 'torque'.}

\item{\code{time_per_zstat}}{The amount of time to budget for each zstat to run through pTFCE in dd-hh:mm:ss format.
  Default is 10:00 (10 minutes).
method to return calls to external ptfce_zstat.R script for each zstat}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_ptfce_calls"></a>}}
\if{latex}{\out{\hypertarget{method-get_ptfce_calls}{}}}
\subsection{Method \code{get_ptfce_calls()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ptfce_spec$get_ptfce_calls(include_complete = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{include_complete}}{if TRUE, return calls for zstats that already appear to have
  pTFCE-corrected images in place. Default: FALSE.
method to submit all ptfce_zstat.R calls to a cluster based on the scheduler specified}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_expected_files"></a>}}
\if{latex}{\out{\hypertarget{method-get_expected_files}{}}}
\subsection{Method \code{get_expected_files()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ptfce_spec$get_expected_files()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-run"></a>}}
\if{latex}{\out{\hypertarget{method-run}{}}}
\subsection{Method \code{run()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ptfce_spec$run(force = FALSE)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-submit"></a>}}
\if{latex}{\out{\hypertarget{method-submit}{}}}
\subsection{Method \code{submit()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ptfce_spec$submit(force = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{force}}{if TRUE, re-run pTFCE for zstat images that already appear to have pTFCE-corrected outputs in place}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-all_expected_exist"></a>}}
\if{latex}{\out{\hypertarget{method-all_expected_exist}{}}}
\subsection{Method \code{all_expected_exist()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ptfce_spec$all_expected_exist()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ptfce_spec$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
