% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_glm_pipeline.R
\name{R_batch_job}
\alias{R_batch_job}
\title{write the sbatch, pbs, or local bash script
write code to be executed}
\description{
write the sbatch, pbs, or local bash script
write code to be executed

write the sbatch, pbs, or local bash script
write code to be executed
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{R_batch_job$new()}}
\item \href{#method-write_batch_file}{\code{R_batch_job$write_batch_file()}}
\item \href{#method-write_compute_file}{\code{R_batch_job$write_compute_file()}}
\item \href{#method-submit}{\code{R_batch_job$submit()}}
\item \href{#method-clone}{\code{R_batch_job$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{R_batch_job$new(
  parent_jobs = NULL,
  job_id = NULL,
  n_nodes = NULL,
  n_cpus = NULL,
  cpu_time = NULL,
  batch_id = NULL,
  code_to_run = NULL,
  r_packages = NULL,
  scheduler = NULL,
  scheduler_args = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-write_batch_file"></a>}}
\if{latex}{\out{\hypertarget{method-write_batch_file}{}}}
\subsection{Method \code{write_batch_file()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{R_batch_job$write_batch_file()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-write_compute_file"></a>}}
\if{latex}{\out{\hypertarget{method-write_compute_file}{}}}
\subsection{Method \code{write_compute_file()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{R_batch_job$write_compute_file()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-submit"></a>}}
\if{latex}{\out{\hypertarget{method-submit}{}}}
\subsection{Method \code{submit()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{R_batch_job$submit()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{R_batch_job$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
