% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_batch_job.R
\name{R_batch_job}
\alias{R_batch_job}
\title{Description of R_batch_job R6 class}
\description{
Description of R_batch_job R6 class

Description of R_batch_job R6 class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{parent_jobs}}{a vector of parent_jobs that are upstream of this job and may influence its execution}

\item{\code{depends_on_parents}}{logical or character string indicating whether this job should wait until
\code{parent_jobs} complete. If a character string is passed, it indicates which named elements of
\code{parent_jobs} are must complete before this job begins.}

\item{\code{child_jobs}}{a vector of job ids generated by the submission of this job. Only relevant if this job
submits additional jobs through the scheduler directly}

\item{\code{job_name}}{a user-defined name for the job used for specifying job dependencies and informative job
status queries on a job scheduler}

\item{\code{cpu_time}}{The amount of time requested on the job scheduler, following d-hh:mm:ss format. Defaults to
"4:00:00", which is 4 hours.}

\item{\code{n_nodes}}{The number of nodes to be requested on the job scheduler}

\item{\code{n_cpus}}{The number of cores (aka 'cpus', ignoring hyperthreading) to be requested on the job scheduler}

\item{\code{mem_total}}{The total amount of memory (RAM) requested by the job}

\item{\code{mem_per_cpu}}{The amount of memory (RAM) requested per cpu (total = mem_per_cpu * n_cpus)}

\item{\code{input_environment}}{The name of the environment to be loaded at the beginning of the R batch prior to executing}

\item{\code{output_environment}}{The name of the environment to be saved at the end of the R batch execution, which can then
be loaded by subsequent jobs.}

\item{\code{sqlite_db}}{Not used yet, but will be used for job tracking in future}

\item{\code{batch_id}}{Not currently used, but intended for job sequence tracking}

\item{\code{batch_directory}}{Location of batch scripts to be written}

\item{\code{batch_code}}{Shell code to be included in the batch script prior to the R code to be run. This can include
module load statements, environment variable exports, etc.}

\item{\code{r_code}}{The R code to be executed by the scheduler. This can be a character vector that includes multiple 
R statements or an expression object containing the R code to be evaluated}

\item{\code{r_packages}}{The R packages to be loaded into the environment before job execution. These are loaded by 
pacman::p_load, which will install any missing packages before attempting to load}

\item{\code{scheduler}}{The job scheduler to be used for this batch. Options are: "slurm", "torque", or "local".}

\item{\code{scheduler_options}}{An optional character vector of scheduler arguments to be included in the batch script header
that control additional features such as job emails or group permissions. These directives are added with #SBATCH
or #PBS headings, depending on the scheduler, and are ignored if the scheduler is "local".}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{R_batch_job$new()}}
\item \href{#method-generate}{\code{R_batch_job$generate()}}
\item \href{#method-submit}{\code{R_batch_job$submit()}}
\item \href{#method-clone}{\code{R_batch_job$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Create a new R_batch_job object for execution on an HPC cluster
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{R_batch_job$new(
  parent_jobs = NULL,
  job_name = NULL,
  n_nodes = NULL,
  n_cpus = NULL,
  cpu_time = NULL,
  mem_per_cpu = NULL,
  mem_total = NULL,
  batch_id = NULL,
  r_code = NULL,
  batch_code,
  r_packages = NULL,
  scheduler = NULL,
  scheduler_options = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{parent_jobs}}{A vector of one or more job ids that are parents to this job. This can be a named vector, to
be used in conjunction with \code{depends_on_parents} to specify which parent jobs must be completed before this
job begins.}

\item{\code{job_name}}{The name of this job}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-generate"></a>}}
\if{latex}{\out{\hypertarget{method-generate}{}}}
\subsection{Method \code{generate()}}{
Helper function that generates the batch and compute files for a job
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{R_batch_job$generate()}\if{html}{\out{</div>}}
}

\subsection{Details}{
this is called by \code{$submit} when a job is submitted and is provided
  here in case the user wants to generate the batch files without executing them
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-submit"></a>}}
\if{latex}{\out{\hypertarget{method-submit}{}}}
\subsection{Method \code{submit()}}{
Submit job to scheduler or local compute
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{R_batch_job$submit()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{R_batch_job$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
