% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_batch_job.R
\name{R_batch_job}
\alias{R_batch_job}
\title{write the sbatch, pbs, or local bash script
write code to be executed
initialization routine for object
Helper function that generates the batch and compute files for a job}
\description{
write the sbatch, pbs, or local bash script
write code to be executed
initialization routine for object
Helper function that generates the batch and compute files for a job

write the sbatch, pbs, or local bash script
write code to be executed
initialization routine for object
Helper function that generates the batch and compute files for a job
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{R_batch_job$new()}}
\item \href{#method-generate}{\code{R_batch_job$generate()}}
\item \href{#method-submit}{\code{R_batch_job$submit()}}
\item \href{#method-clone}{\code{R_batch_job$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{R_batch_job$new(
  parent_jobs = NULL,
  job_id = NULL,
  job_name = NULL,
  n_nodes = NULL,
  n_cpus = NULL,
  cpu_time = NULL,
  mem_per_cpu = NULL,
  mem_total = NULL,
  batch_id = NULL,
  r_code = NULL,
  batch_code,
  r_packages = NULL,
  scheduler = NULL,
  scheduler_options = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-generate"></a>}}
\if{latex}{\out{\hypertarget{method-generate}{}}}
\subsection{Method \code{generate()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{R_batch_job$generate()}\if{html}{\out{</div>}}
}

\subsection{Details}{
this is called by \code{$submit} when a job is submitted and is provided
  here in case the user wants to generate the batch files without executing them
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-submit"></a>}}
\if{latex}{\out{\hypertarget{method-submit}{}}}
\subsection{Method \code{submit()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{R_batch_job$submit()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{R_batch_job$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
