% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_batch_job.R
\name{R_batch_sequence}
\alias{R_batch_sequence}
\title{Description of R_batch_sequence R6 class}
\description{
Description of R_batch_sequence R6 class

Description of R_batch_sequence R6 class
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{R_batch_sequence$new()}}
\item \href{#method-submit}{\code{R_batch_sequence$submit()}}
\item \href{#method-clone}{\code{R_batch_sequence$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
create a new R_batch_sequence object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{R_batch_sequence$new(..., joblist = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-submit"></a>}}
\if{latex}{\out{\hypertarget{method-submit}{}}}
\subsection{Method \code{submit()}}{
submit the job sequence to the scheduler or local compute
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{R_batch_sequence$submit()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{R_batch_sequence$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
