% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_df_sqlite.R
\name{read_df_sqlite}
\alias{read_df_sqlite}
\title{helper function to lookup a keyed data.frame from the sqlite storage database}
\usage{
read_df_sqlite(
  gpa = NULL,
  db_file = NULL,
  id = NULL,
  session = NULL,
  run_number = NULL,
  table = NULL,
  drop_keys = TRUE,
  quiet = TRUE
)
}
\arguments{
\item{gpa}{A \code{glm_pipeline_arguments} object used to lookup location of SQLite database for this analysis}

\item{db_file}{An optional string specifying the SQLite database from which to read}

\item{id}{the id of the subject to whom these data belong}

\item{session}{the session of these data}

\item{run_number}{the run_number of these data}

\item{table}{A character string of the table name from which to read}

\item{drop_keys}{whether to drop identifying metatdata columns from data before returning the object}

\item{quiet}{a logical indicating whether to issue a warning if the table is not found}
}
\value{
a data.frame containing the requested data. Will return NULL if not found
}
\description{
helper function to lookup a keyed data.frame from the sqlite storage database
}
\keyword{internal}
