% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glm_helper_functions.R
\name{run_afni_command}
\alias{run_afni_command}
\title{Wrapper for running an AFNI command safely within R}
\usage{
run_afni_command(
  args,
  afnidir = NULL,
  stdout = NULL,
  stderr = NULL,
  echo = TRUE,
  ...
)
}
\arguments{
\item{args}{AFNI command string to be run}

\item{afnidir}{Location of AFNI installation. If NULL, the function will serach the environment for AFNIDIF}

\item{stdout}{File target for redirecting stdout. If NULL, stdout will not be captured}

\item{stderr}{File target for redirecting stderr. If NULL, stderr will not be captured}

\item{echo}{Whether to print AFNI command to the screen}
}
\value{
The exit status of the executed AFNI command. 0 for success, non-zero for failure
}
\description{
Wrapper for running an AFNI command safely within R
}
\details{
This command ensures that AFNI comamnds are in the system PATH
}
\examples{

\dontrun{
runAFNICommand("3dcopy test_data copy_data")
}
}
\author{
Michael Hallquist
}
