% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/afni_3dlmer.R
\name{sanitize_3dlmer_model_formula}
\alias{sanitize_3dlmer_model_formula}
\title{Internal function to construct the 3dLMEr command string}
\usage{
sanitize_3dlmer_model_formula(model_formula)
}
\arguments{
\item{model_formula}{fixed effects formula string}
}
\value{
a compact 3dLMEr formula string
}
\description{
Internal function to construct the 3dLMEr command string
}
\keyword{internal}
