% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specify_contrasts.R
\name{specify_contrasts}
\alias{specify_contrasts}
\title{Internal function to add, edit, rename, delete contrasts}
\usage{
specify_contrasts(
  mobj = NULL,
  include_diagonal = TRUE,
  include_factormeans = FALSE
)
}
\arguments{
\item{mobj}{\code{l1_model_spec} or \code{hi_model_spec} object}

\item{include_diagonal}{A logical indicating whether to include a diagonal matrix of contrasts
in the model contrasts. Defaults to TRUE.}
}
\description{
Internal function to add, edit, rename, delete contrasts
}
\author{
Michael Hallquist
}
