% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sqlite_functions.R
\name{submit_sqlite_query}
\alias{submit_sqlite_query}
\title{helper function to establish sqlite connection and submit query}
\usage{
submit_sqlite_query(
  str = NULL,
  sqlite_db = NULL,
  param = NULL,
  busy_timeout = NULL,
  return_result = FALSE
)
}
\arguments{
\item{str}{Character query statement to execute}

\item{sqlite_db}{Character path to SQLite database}

\item{busy_timeout}{Time (in s) after which to retry write operations; default is 10 s}

\item{return_result}{Logical. If TRUE submits DBI::dbGetQuery instead of DBI::dbExecute;
Only use if expecting something in return for your query}

\item{params}{Optional list of parameters to pass to statement}
}
\description{
helper function to establish sqlite connection and submit query
}
\keyword{internal}
