% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_glm_pipeline.R
\name{summary.bg_subject_data}
\alias{summary.bg_subject_data}
\title{Summarize the contents of the subject data}
\usage{
\method{summary}{bg_subject_data}(df)
}
\arguments{
\item{df}{A \code{bg_subject_data} object containing subject-level data.}
}
\description{
Summarize the contents of the subject data
}
\details{
This is typically run using \code{summary(gpa$trial_data)}
}
