% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_funcs.R
\name{summary.glm_pipeline_arguments}
\alias{summary.glm_pipeline_arguments}
\title{Helper function to better summarize GPA object}
\usage{
\method{summary}{glm_pipeline_arguments}(gpa)
}
\arguments{
\item{gpa}{the \code{gpa} object}
}
\value{
a readable summary of the given gpa object
}
\description{
Helper function to better summarize GPA object
}
