% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_functions.R
\name{summary.l1_model_set}
\alias{summary.l1_model_set}
\title{Helper function to better summarize GPA object's L1 models}
\usage{
\method{summary}{l1_model_set}(l1_ms)
}
\arguments{
\item{gpa}{the \code{gpa} object}
}
\value{
a readable summary of the given gpa object's L1 models
}
\description{
Helper function to better summarize GPA object's L1 models
}
