% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_functions.R
\name{summary.l1_model_set_events}
\alias{summary.l1_model_set_events}
\title{Helper function to better summarize GPA object's l1_models' events}
\usage{
\method{summary}{l1_model_set_events}(l1_model_set_events)
}
\arguments{
\item{gpa}{the \code{gpa} object}
}
\value{
a readable summary of the given gpa object's l1_models' events
}
\description{
Helper function to better summarize GPA object's l1_models' events
}
