% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_l1_confounds.R
\name{test_exclude_run}
\alias{test_exclude_run}
\title{calculate whether to retain or exclude this run}
\usage{
test_exclude_run(
  gpa,
  id,
  session,
  run_number,
  generate_run_exclusion,
  all_confounds,
  lg,
  confound_path = NA_character_,
  motion_path = NA_character_
)
}
\description{
calculate whether to retain or exclude this run
}
