% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/truncate_runs.R
\name{truncate_runs}
\alias{truncate_runs}
\title{Identify the last valid volume acquired in a given run.
Subjects often exhibit head movement after run ends (MATLAB closes), but scan hasn't stopped
This occurs because the MB raw transfer of the prior run is occurring, but does not finish before the current run
Thus, truncate mr files to be 12 seconds after final feedback presentation, which is how the paradigm timing files are setup}
\usage{
truncate_runs(mr_df, subj_outdir = NULL, lg = NULL)
}
\description{
Identify the last valid volume acquired in a given run.
Subjects often exhibit head movement after run ends (MATLAB closes), but scan hasn't stopped
This occurs because the MB raw transfer of the prior run is occurring, but does not finish before the current run
Thus, truncate mr files to be 12 seconds after final feedback presentation, which is how the paradigm timing files are setup
}
\keyword{internal}
